Fantasy draft history

Return the sequential result of a fantasy draft pick, whether snake or salary cap format.

draft_recap(leagueId = ffl_id(), leagueHistory = FALSE, ...)

Arguments

leagueId
Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

leagueHistory
logical; Should the leagueHistory version of the API be called? If TRUE, a list of results is returned, with one element for each historical year of the league.

...
Additional queries passed to httr::GET() (e.g., scoringPeriodId). Arguments are converted to a named list to be passed alongside view.

Value

A data frame(s) of draft picks.

Examples

draft_recap(leagueId = "42654852")

#> # A tibble: 64 × 14
#>    seasonId autoDraftTypeId bidAmount    id keeper lineupSlotId nominatingTeamId
#>       <int> <lgl>               <int> <int> <lgl>         <int> <fct>           
#>  1     2021 TRUE                   NA     1 FALSE             2 NA              
#>  2     2021 TRUE                   NA     2 FALSE             2 NA              
#>  3     2021 TRUE                   NA     3 FALSE             2 NA              
#>  4     2021 TRUE                   NA     4 FALSE             6 NA              
#>  5     2021 TRUE                   NA     5 FALSE             2 NA              
#>  6     2021 TRUE                   NA     6 FALSE             2 NA              
#>  7     2021 TRUE                   NA     7 FALSE             2 NA              
#>  8     2021 TRUE                   NA     8 FALSE             2 NA              
#>  9     2021 TRUE                   NA     9 FALSE             4 NA              
#> 10     2021 TRUE                   NA    10 FALSE             4 NA              
#> # … with 54 more rows, and 7 more variables: overallPickNumber <int>,
#> #   playerId <int>, reservedForKeeper <lgl>, roundId <int>,
#> #   roundPickNumber <int>, teamId <fct>, tradeLocked <lgl>

draft_recap(leagueId = "252353", leagueHistory = TRUE)

#> [[1]]
#> # A tibble: 128 × 16
#>    seasonId autoDraftTypeId bidAmount    id keeper lineupSlotId memberId        
#>       <int> <lgl>               <int> <int> <lgl>         <int> <chr>           
#>  1     2015 FALSE                  NA     1 FALSE             2 3436600c-3252-4…
#>  2     2015 FALSE                  NA     2 FALSE             2 3436600c-3252-4…
#>  3     2015 FALSE                  NA     3 FALSE             2 3436600c-3252-4…
#>  4     2015 FALSE                  NA     4 FALSE             2 3436600c-3252-4…
#>  5     2015 FALSE                  NA     5 FALSE             2 3436600c-3252-4…
#>  6     2015 FALSE                  NA     6 FALSE             2 3436600c-3252-4…
#>  7     2015 FALSE                  NA     7 FALSE             4 3436600c-3252-4…
#>  8     2015 FALSE                  NA     8 FALSE             2 3436600c-3252-4…
#>  9     2015 FALSE                  NA     9 FALSE             6 3436600c-3252-4…
#> 10     2015 FALSE                  NA    10 FALSE             4 3436600c-3252-4…
#> # … with 118 more rows, and 9 more variables: nominatingTeamId <fct>,
#> #   overallPickNumber <int>, owningTeamIds <list>, playerId <int>,
#> #   reservedForKeeper <lgl>, roundId <int>, roundPickNumber <int>,
#> #   teamId <fct>, tradeLocked <lgl>
#> 
#> [[2]]
#> # A tibble: 160 × 16
#>    seasonId autoDraftTypeId bidAmount    id keeper lineupSlotId memberId        
#>       <int> <lgl>               <int> <int> <lgl>         <int> <chr>           
#>  1     2016 FALSE                  NA     1 FALSE             4 3436600c-3252-4…
#>  2     2016 FALSE                  NA     2 FALSE             4 3436600c-3252-4…
#>  3     2016 FALSE                  NA     3 FALSE             4 3436600c-3252-4…
#>  4     2016 FALSE                  NA     4 FALSE             2 3436600c-3252-4…
#>  5     2016 FALSE                  NA     5 FALSE             2 3436600c-3252-4…
#>  6     2016 FALSE                  NA     6 FALSE             2 3436600c-3252-4…
#>  7     2016 FALSE                  NA     7 FALSE             4 3436600c-3252-4…
#>  8     2016 FALSE                  NA     8 FALSE             6 3436600c-3252-4…
#>  9     2016 FALSE                  NA     9 FALSE             2 3436600c-3252-4…
#> 10     2016 FALSE                  NA    10 FALSE             4 3436600c-3252-4…
#> # … with 150 more rows, and 9 more variables: nominatingTeamId <fct>,
#> #   overallPickNumber <int>, owningTeamIds <list>, playerId <int>,
#> #   reservedForKeeper <lgl>, roundId <int>, roundPickNumber <int>,
#> #   teamId <fct>, tradeLocked <lgl>
#> 
#> [[3]]
#> # A tibble: 160 × 16
#>    seasonId autoDraftTypeId bidAmount    id keeper lineupSlotId memberId        
#>       <int> <lgl>               <int> <int> <lgl>         <int> <chr>           
#>  1     2017 FALSE                  28     1 FALSE             0 3436600c-3252-4…
#>  2     2017 FALSE                  25     2 FALSE             0 3436600c-3252-4…
#>  3     2017 FALSE                  67     3 FALSE             2 3436600c-3252-4…
#>  4     2017 FALSE                  56     4 FALSE             2 3436600c-3252-4…
#>  5     2017 FALSE                  15     5 FALSE             4 3436600c-3252-4…
#>  6     2017 FALSE                  49     6 FALSE             2 3436600c-3252-4…
#>  7     2017 FALSE                  75     7 FALSE             2 3436600c-3252-4…
#>  8     2017 FALSE                  53     8 FALSE             4 3436600c-3252-4…
#>  9     2017 FALSE                  46     9 FALSE             2 3436600c-3252-4…
#> 10     2017 FALSE                  42    10 FALSE             2 3436600c-3252-4…
#> # … with 150 more rows, and 9 more variables: nominatingTeamId <fct>,
#> #   overallPickNumber <int>, owningTeamIds <list>, playerId <int>,
#> #   reservedForKeeper <lgl>, roundId <int>, roundPickNumber <int>,
#> #   teamId <fct>, tradeLocked <lgl>
#> 
#> [[4]]
#> # A tibble: 160 × 16
#>    seasonId autoDraftTypeId bidAmount    id keeper lineupSlotId memberId        
#>       <int> <lgl>               <int> <int> <lgl>         <int> <chr>           
#>  1     2018 FALSE                  37     1 FALSE             2 2c6bd94c-3684-4…
#>  2     2018 FALSE                  19     2 FALSE             4 2c6bd94c-3684-4…
#>  3     2018 FALSE                  19     3 FALSE             4 2c6bd94c-3684-4…
#>  4     2018 FALSE                  33     4 FALSE             2 2c6bd94c-3684-4…
#>  5     2018 FALSE                  65     5 FALSE             2 2c6bd94c-3684-4…
#>  6     2018 FALSE                  45     6 FALSE             0 2c6bd94c-3684-4…
#>  7     2018 FALSE                  42     7 FALSE             6 2c6bd94c-3684-4…
#>  8     2018 FALSE                  51     8 FALSE             2 2c6bd94c-3684-4…
#>  9     2018 FALSE                  68     9 FALSE             2 2c6bd94c-3684-4…
#> 10     2018 FALSE                  72    10 FALSE             2 2c6bd94c-3684-4…
#> # … with 150 more rows, and 9 more variables: nominatingTeamId <fct>,
#> #   overallPickNumber <int>, owningTeamIds <list>, playerId <int>,
#> #   reservedForKeeper <lgl>, roundId <int>, roundPickNumber <int>,
#> #   teamId <fct>, tradeLocked <lgl>
#> 
#> [[5]]
#> # A tibble: 128 × 15
#>    seasonId autoDraftTypeId bidAmount    id keeper lineupSlotId memberId        
#>       <int> <lgl>               <int> <int> <lgl>         <int> <chr>           
#>  1     2019 FALSE                   3     1 FALSE            17 {35D9B7DB-E821-…
#>  2     2019 FALSE                  21     2 FALSE             4 {9A722A48-6E16-…
#>  3     2019 FALSE                  43     3 FALSE             2 {12E61406-349A-…
#>  4     2019 FALSE                  35     4 FALSE             2 {9A722A48-6E16-…
#>  5     2019 FALSE                  56     5 FALSE             4 {F33F0723-2CC3-…
#>  6     2019 FALSE                  43     6 FALSE             2 {91F8B424-1689-…
#>  7     2019 FALSE                  48     7 FALSE             4 {0D279541-6015-…
#>  8     2019 FALSE                  23     8 FALSE             2 {12E61406-349A-…
#>  9     2019 FALSE                  41     9 FALSE             4 {5FA14794-6573-…
#> 10     2019 FALSE                  37    10 FALSE             2 {91F8B424-1689-…
#> # … with 118 more rows, and 8 more variables: nominatingTeamId <fct>,
#> #   overallPickNumber <int>, playerId <int>, reservedForKeeper <lgl>,
#> #   roundId <int>, roundPickNumber <int>, teamId <fct>, tradeLocked <lgl>
#> 
#> [[6]]
#> # A tibble: 128 × 15
#>    seasonId autoDraftTypeId bidAmount    id keeper lineupSlotId memberId        
#>       <int> <lgl>               <int> <int> <lgl>         <int> <chr>           
#>  1     2020 FALSE                  57     1 FALSE             2 {91F8B424-1689-…
#>  2     2020 FALSE                  33     2 FALSE             0 {9A722A48-6E16-…
#>  3     2020 FALSE                  23     3 FALSE             4 {5FA14794-6573-…
#>  4     2020 FALSE                   3     4 FALSE             0 {12E61406-349A-…
#>  5     2020 FALSE                  32     5 FALSE             2 {12E61406-349A-…
#>  6     2020 FALSE                  81     6 FALSE             2 {9A722A48-6E16-…
#>  7     2020 FALSE                  45     7 FALSE             2 {F33F0723-2CC3-…
#>  8     2020 FALSE                   7     8 FALSE             6 {35D9B7DB-E821-…
#>  9     2020 FALSE                  52     9 FALSE             2 {35D9B7DB-E821-…
#> 10     2020 FALSE                  32    10 FALSE             4 {C61F403A-F689-…
#> # … with 118 more rows, and 8 more variables: nominatingTeamId <fct>,
#> #   overallPickNumber <int>, playerId <int>, reservedForKeeper <lgl>,
#> #   roundId <int>, roundPickNumber <int>, teamId <fct>, tradeLocked <lgl>
#>