League draft settings

The type, date, and pick order of a league draft.

draft_settings(leagueId = ffl_id(), leagueHistory = FALSE, ...)

Arguments

leagueId
Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

leagueHistory
logical; Should the leagueHistory version of the API be called? If TRUE, a list of results is returned, with one element for each historical year of the league.

...
Additional queries passed to httr::GET() (e.g., scoringPeriodId). Arguments are converted to a named list to be passed alongside view.

Value

A data frame of league draft settings by season.

See also

Examples

draft_settings(leagueId = "42654852")

#> # A tibble: 1 × 13
#>   seasonId auctionBudget availableDate       date               
#>      <int> <chr>         <dttm>              <dttm>             
#> 1     2021 NA            2021-08-03 14:15:00 2021-08-03 15:15:00
#> # … with 9 more variables: isTradingEnabled <lgl>, keeperCount <int>,
#> #   keeperCountFuture <int>, keeperOrderType <chr>, leagueSubType <chr>,
#> #   orderType <chr>, pickOrder <list>, timePerSelection <int>, type <chr>

draft_settings(leagueId = "252353", leagueHistory = TRUE)

#> # A tibble: 6 × 13
#>   seasonId auctionBudget availableDate       date               
#>      <int> <chr>         <dttm>              <dttm>             
#> 1     2015 NA            2015-08-15 16:00:00 2015-08-15 16:00:00
#> 2     2016 NA            2016-09-05 11:30:00 2016-09-05 11:30:00
#> 3     2017 200           2017-08-31 18:00:00 2017-08-31 18:00:00
#> 4     2018 200           2018-09-05 18:00:00 2018-09-05 18:00:00
#> 5     2019 200           2019-09-01 23:00:00 2019-09-02 00:00:00
#> 6     2020 200           2020-09-06 23:15:00 2020-09-07 00:15:00
#> # … with 9 more variables: isTradingEnabled <lgl>, keeperCount <int>,
#> #   keeperCountFuture <int>, keeperOrderType <chr>, leagueSubType <chr>,
#> #   orderType <chr>, pickOrder <list>, timePerSelection <int>, type <chr>