Find fantasy players

Filter fantasy players by their position, availability, professional team, and/or injury status. Sort and limit the responses in the same way as is done in the ESPN Fantasy Football website.

list_players(
  leagueId = ffl_id(),
  sort = "ROST",
  position = NULL,
  status = "AVAILABLE",
  injured = NULL,
  proTeam = NULL,
  scoreType = c("STANDARD", "PPR"),
  limit = 50
)

Arguments

leagueId

Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

sort

The column from which to sort the data. Options match those on the ESPN website:

"PLAYER" = Alphabetical by player name

"PROJ" = Projection is ESPN’s projected fantasy score for a player’s upcoming game.

"SCORE" = Actual score for scoringPeriodId

"OPRK" = Opponent Rank shows how a player’s upcoming NFL opponent performs against that player’s position. Low numbers mean it may be a tough opponent; high numbers an easier opponent.

"START" = Start Percentage shows the number of fantasy leagues a player is started in divided by the number of leagues he is eligible in. This helps indicate how the public views a player.

"ROST" = Rostered Percentage shows the number of fantasy leagues in which a player is on a roster divided by the total number of fantasy leagues. This helps indicate how the public views a player.

"CHANGE" = Plus/Minus shows the change in %ROST over the last week. This will help show which players are hot and cold at a given moment.

"PRK" = Position Rank shows how a player stacks up against other players at his position. No. 1 is best.

"FPTS" = Total fantasy points scored thus far in the season.

"AVG" = Average fantasy points scored in each game started.

"LAST" = Last shows the player’s fantasy score in his team’s last game.

position

Abbreviation of player positions to filter, NULL for all:

"QB" = Quarterback

"RB" = Running Back

"WR" = Wide Receiver

"TE" = Tight End

"FLEX" = Running Backs, Wide Receivers and Tight Ends can be used in this position

"D/ST" = Defense and Special Teams

"K" = Kicker

status

Availability status of player, one or more from:

"ALL"

"AVAILABLE" (default)

"FREEAGENT"

"WAIVERS"

"ONTEAM"

injured

Whether to return only injured or healthy players. Use NULL (default) for all players, TRUE for injured players, and FALSE for healthy players.

proTeam

The abbreviation or ID of the professional team from which players should be returned. See pro_teams() for a list of all possible team abbreviations.

scoreType

The type of scoring used: "STANDARD" or "PPR."

limit

The limit of players to return. Use "" or NULL to return all. Defaults to 50, which is the default limit used by ESPN. Removing the limit can make the request take a long time.

Value

A data frame of players.

See also

Examples

list_players("42654852", proTeam = "Mia", sort = "START", limit = 3)
#> # A tibble: 3 × 19
#>   seasonId scoringPeriodId      id firstName lastName   proTeam defaultPosition
#>      <int>           <dbl>   <int> <chr>     <chr>      <fct>   <fct>          
#> 1     2023              19 4429160 De'Von    Achane     Mia     RB             
#> 2     2023              19  -16015 Dolphins  D/ST       Mia     D/ST           
#> 3     2023              19 4241479 Tua       Tagovailoa Mia     QB             
#> # ℹ 12 more variables: injuryStatus <chr>, percentStarted <dbl>,
#> #   percentOwned <dbl>, percentChange <dbl>, positionalRanking <int>,
#> #   totalRating <dbl>, auctionValueAverage <dbl>, averageDraftPosition <dbl>,
#> #   projectedScore <dbl>, lastScore <dbl>, lastSeason <dbl>,
#> #   currentSeason <dbl>