Live matchup scoreboard

The current and projected score for each ongoing match.

live_scoring(lid = getOption("lid"), yet = FALSE)

Arguments

lid
ESPN League ID, defaults to getOption("lid") (see set_lid()).

yet
If TRUE, pro_schedule() and the "mRoster" view are called to determine how many starting players have yet started playing.

Value

A tibble of scores.

Examples

live_scoring(252353)

#> # A tibble: 8 x 5
#>    week match team  score  proj
#>   <int> <int> <fct> <dbl> <dbl>
#> 1     6    21 AGUS   79.3 103. 
#> 2     6    21 CART   57.5  70.0
#> 3     6    22 NICK  131.  139. 
#> 4     6    22 CORE   49.5  68.1
#> 5     6    23 PEPE   21.9  79.5
#> 6     6    23 KIER   53.4  77.1
#> 7     6    24 BILL   86.3  86.3
#> 8     6    24 KYLE   44.5  97.8