Fantasy match schedule

The opponents each team faces every week in a fantasy regular season. Returned in a tidy format where each row is a single team with an indication of home-away status. There are two rows per matchup, one for each team.

tidy_matchups(leagueId = ffl_id(), leagueHistory = FALSE, ...)

Arguments

leagueId
Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

leagueHistory
logical; Should the leagueHistory version of the API be called? If TRUE, a list of results is returned, with one element for each historical year of the league.

...
Additional queries passed to httr::GET() (e.g., scoringPeriodId). Arguments are converted to a named list to be passed alongside view.

Value

A data frame(s) of match opponents.

Examples

tidy_matchups(leagueId = "42654852")

#> # A tibble: 68 × 6
#>    seasonId matchupPeriodId    id teamId opponent home 
#>       <int>           <int> <int> <fct>  <fct>    <lgl>
#>  1     2021               1     1 AUS    CHI      TRUE 
#>  2     2021               1     1 CHI    AUS      FALSE
#>  3     2021               1     2 BOS    DEN      TRUE 
#>  4     2021               1     2 DEN    BOS      FALSE
#>  5     2021               2     3 CHI    DEN      TRUE 
#>  6     2021               2     3 DEN    CHI      FALSE
#>  7     2021               2     4 AUS    BOS      TRUE 
#>  8     2021               2     4 BOS    AUS      FALSE
#>  9     2021               3     5 DEN    AUS      TRUE 
#> 10     2021               3     5 AUS    DEN      FALSE
#> # … with 58 more rows

tidy_matchups(leagueId = "252353", leagueHistory = TRUE)

#> [[1]]
#> # A tibble: 120 × 6
#>    seasonId matchupPeriodId    id teamId opponent home 
#>       <int>           <int> <int> <fct>  <fct>    <lgl>
#>  1     2015               1     0 GRIZ   BILL     TRUE 
#>  2     2015               1     0 BILL   GRIZ     FALSE
#>  3     2015               1     1 BREW   CHAR     TRUE 
#>  4     2015               1     1 CHAR   BREW     FALSE
#>  5     2015               1     2 LOVE   GUS      TRUE 
#>  6     2015               1     2 GUS    LOVE     FALSE
#>  7     2015               1     3 K5     PEPE     TRUE 
#>  8     2015               1     3 PEPE   K5       FALSE
#>  9     2015               2     4 BILL   CHAR     TRUE 
#> 10     2015               2     4 CHAR   BILL     FALSE
#> # … with 110 more rows
#> 
#> [[2]]
#> # A tibble: 150 × 6
#>    seasonId matchupPeriodId    id teamId opponent home 
#>       <int>           <int> <int> <fct>  <fct>    <lgl>
#>  1     2016               1     0 PEPE   MOOD     TRUE 
#>  2     2016               1     0 MOOD   PEPE     FALSE
#>  3     2016               1     1 CHAR   NICK     TRUE 
#>  4     2016               1     1 NICK   CHAR     FALSE
#>  5     2016               1     2 TEAM   K5       TRUE 
#>  6     2016               1     2 K5     TEAM     FALSE
#>  7     2016               1     3 BILL   GUS      TRUE 
#>  8     2016               1     3 GUS    BILL     FALSE
#>  9     2016               1     4 LOVE   BRUZ     TRUE 
#> 10     2016               1     4 BRUZ   LOVE     FALSE
#> # … with 140 more rows
#> 
#> [[3]]
#> # A tibble: 140 × 6
#>    seasonId matchupPeriodId    id teamId opponent home 
#>       <int>           <int> <int> <fct>  <fct>    <lgl>
#>  1     2017               1     0 PEPE   MOOD     TRUE 
#>  2     2017               1     0 MOOD   PEPE     FALSE
#>  3     2017               1     1 CHAR   NICK     TRUE 
#>  4     2017               1     1 NICK   CHAR     FALSE
#>  5     2017               1     2 COLN   KIER     TRUE 
#>  6     2017               1     2 KIER   COLN     FALSE
#>  7     2017               1     3 BILL   GUS      TRUE 
#>  8     2017               1     3 GUS    BILL     FALSE
#>  9     2017               1     4 LOVE   CART     TRUE 
#> 10     2017               1     4 CART   LOVE     FALSE
#> # … with 130 more rows
#> 
#> [[4]]
#> # A tibble: 140 × 6
#>    seasonId matchupPeriodId    id teamId opponent home 
#>       <int>           <int> <int> <fct>  <fct>    <lgl>
#>  1     2018               1     0 PEPE   KYLE     TRUE 
#>  2     2018               1     0 KYLE   PEPE     FALSE
#>  3     2018               1     1 CHAR   NICK     TRUE 
#>  4     2018               1     1 NICK   CHAR     FALSE
#>  5     2018               1     2 COLN   KIER     TRUE 
#>  6     2018               1     2 KIER   COLN     FALSE
#>  7     2018               1     3 BILL   GUS      TRUE 
#>  8     2018               1     3 GUS    BILL     FALSE
#>  9     2018               1     4 ROWN   CART     TRUE 
#> 10     2018               1     4 CART   ROWN     FALSE
#> # … with 130 more rows
#> 
#> [[5]]
#> # A tibble: 112 × 6
#>    seasonId matchupPeriodId    id teamId opponent home 
#>       <int>           <int> <int> <fct>  <fct>    <lgl>
#>  1     2019               1     1 CORY   BILL     TRUE 
#>  2     2019               1     1 BILL   CORY     FALSE
#>  3     2019               1     2 NICK   KIER     TRUE 
#>  4     2019               1     2 KIER   NICK     FALSE
#>  5     2019               1     3 KYLE   CART     TRUE 
#>  6     2019               1     3 CART   KYLE     FALSE
#>  7     2019               1     4 PEPE   ANGS     TRUE 
#>  8     2019               1     4 ANGS   PEPE     FALSE
#>  9     2019               2     5 BILL   KIER     TRUE 
#> 10     2019               2     5 KIER   BILL     FALSE
#> # … with 102 more rows
#> 
#> [[6]]
#> # A tibble: 112 × 6
#>    seasonId matchupPeriodId    id teamId opponent home 
#>       <int>           <int> <int> <fct>  <fct>    <lgl>
#>  1     2020               1     1 PEPE   AGUS     TRUE 
#>  2     2020               1     1 AGUS   PEPE     FALSE
#>  3     2020               1     2 NICK   BILL     TRUE 
#>  4     2020               1     2 BILL   NICK     FALSE
#>  5     2020               1     3 CART   KYLE     TRUE 
#>  6     2020               1     3 KYLE   CART     FALSE
#>  7     2020               1     4 CORE   KIER     TRUE 
#>  8     2020               1     4 KIER   CORE     FALSE
#>  9     2020               2     5 AGUS   BILL     TRUE 
#> 10     2020               2     5 BILL   AGUS     FALSE
#> # … with 102 more rows
#>