Fantasy matchup scores

The score of each team in a matchup or scoring period and the match outcome.

tidy_scores(leagueId = ffl_id(), leagueHistory = FALSE, useMatchup = TRUE, ...)

Arguments

leagueId
Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

leagueHistory
logical; Should the leagueHistory version of the API be called? If TRUE, a list of results is returned, with one element for each historical year of the league.

useMatchup
logical; Whether scoring should be summarized by matchupPeriodId (default) or scoringPeriodId. The later always relates to a single week of the NFL season, while fantasy matchups might span several scoring periods, especially in the playoffs.

...
Additional queries passed to httr::GET(). Arguments are converted to a named list and passed to query alongside view.

Value

A tidy data frame of scores by team and matchup/scoring period.

Details

expectedWins are calculated by comparing a team score against all other scores for a given matchup period. This statistic expresses how a team would fair if the schedule was random The highest scoring team is thus expected to earn 1 win and the lowest scoring team would expect to win 0 matchups.

See also

Examples

tidy_scores(leagueId = "42654852", useMatchup = FALSE)
#> # A tibble: 32 × 8
#>    seasonId scoringPeriodId matchupId teamId abbrev isHome points expectedWins
#>       <int>           <dbl>     <int>  <int> <fct>  <lgl>   <dbl>        <dbl>
#>  1     2021               1         1      1 AUS    TRUE    117.         0    
#>  2     2021               1         1      3 CHI    FALSE   119.         0.333
#>  3     2021               1         2      2 BOS    TRUE    163.         1    
#>  4     2021               1         2      4 DEN    FALSE   135.         0.667
#>  5     2021               2         3      3 CHI    TRUE    154.         0.667
#>  6     2021               2         3      4 DEN    FALSE   160.         1    
#>  7     2021               2         4      1 AUS    TRUE    139.         0.333
#>  8     2021               2         4      2 BOS    FALSE    93.7        0    
#>  9     2021               3         5      4 DEN    TRUE    127.         0.333
#> 10     2021               3         5      1 AUS    FALSE   118.         0    
#> # … with 22 more rows