League trade settings

The time each trade can stand, votes needed to veto, and season deadline.

trade_settings(leagueId = ffl_id(), leagueHistory = FALSE, ...)

Arguments

leagueId

Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

leagueHistory

logical; Should the leagueHistory version of the API be called? If TRUE, a list of results is returned, with one element for each historical year of the league.

...

Additional queries passed to httr::GET(). Arguments are converted to a named list and passed to query alongside view.

Value

A data frame of league trade settings by season.

See also

Examples

trade_settings(leagueId = "42654852")
#> # A tibble: 1 × 6
#>   seasonId allowOutOfUniverse deadlineDate          max revisionHours vetoVote…¹
#>      <int> <lgl>              <dttm>              <int>         <int>      <int>
#> 1     2022 FALSE              2022-11-30 17:00:00    -1            48          1
#> # … with abbreviated variable name ¹​vetoVotesRequired