League transactions

Summary of transactions and roster changes made during a season by team.

transaction_counter(leagueId = ffl_id(), leagueHistory = FALSE, ...)

budget_summary(leagueId = ffl_id(), leagueHistory = FALSE, ...)

Arguments

leagueId
Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

leagueHistory
logical; Should the leagueHistory version of the API be called? If TRUE, a list of results is returned, with one element for each historical year of the league.

...
Additional queries passed to httr::GET(). Arguments are converted to a named list and passed to query alongside view.

Value

A data frame of transaction counts by team.

See also

Examples

transaction_counter(leagueId = "42654852")
#> # A tibble: 4 × 14
#>   seasonId scoringPeriodId teamId abbrev waiverRank acquisitionBudgetSpent
#>      <int>           <int>  <int> <fct>       <int>                  <int>
#> 1     2021              11      1 AUS             2                      0
#> 2     2021              11      2 BOS             1                      0
#> 3     2021              11      3 CHI             4                      0
#> 4     2021              11      4 DEN             3                      0
#> # … with 8 more variables: acquisitions <int>, drops <int>, misc <int>,
#> #   moveToActive <int>, moveToIR <int>, paid <dbl>, teamCharges <dbl>,
#> #   trades <int>