Sort the optimal fantasy roster

Uses the roster settings for each league to find the best possible combinations of players to score the most fantasy points.

best_roster(
  leagueId = ffl_id(),
  useScore = c("actualScore", "projectedScore"),
  scoringPeriodId = NULL,
  ...
)

Arguments

leagueId

Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

useScore

One of "projectedScore" or "actualScore" (default).

scoringPeriodId

Integer week of NFL season. By default, NULL will use the current week (see ffl_week()). Scoring periods are always one week in length, whereas matchups might be longer.

...

Additional queries passed to httr::GET(). Arguments are converted to a named list and passed to query alongside view.

Value

A dataframe (or list) with optimal rosters.

Details

If scoringPeriodId is the current week (the default), then actual scoring might be incomplete (see projectedScore argument).

See also

Examples

best_roster(leagueId = "42654852", scoringPeriodId = 1)
#> $AUS
#> # A tibble: 16 × 17
#>    seasonId scoringPeriodId teamId abbrev lineupSlot actualSlot playerId
#>       <int>           <int>  <int> <fct>  <fct>      <fct>         <int>
#>  1     2023               1      1 AUS    QB         QB          3139477
#>  2     2023               1      1 AUS    RB         RB          3117251
#>  3     2023               1      1 AUS    RB         FLEX        3916148
#>  4     2023               1      1 AUS    WR         BE          4361370
#>  5     2023               1      1 AUS    WR         BE          4241478
#>  6     2023               1      1 AUS    TE         TE          2576925
#>  7     2023               1      1 AUS    FLEX       RB          4430807
#>  8     2023               1      1 AUS    D/ST       D/ST         -16018
#>  9     2023               1      1 AUS    K          K             15683
#> 10     2023               1      1 AUS    BE         BE            15795
#> 11     2023               1      1 AUS    BE         WR          4372016
#> 12     2023               1      1 AUS    BE         BE          3126486
#> 13     2023               1      1 AUS    BE         BE          4567048
#> 14     2023               1      1 AUS    BE         BE          3116165
#> 15     2023               1      1 AUS    BE         BE          4360238
#> 16     2023               1      1 AUS    BE         WR          2977187
#> # ℹ 10 more variables: firstName <chr>, lastName <chr>, proTeam <fct>,
#> #   position <fct>, injuryStatus <chr>, projectedScore <dbl>,
#> #   actualScore <dbl>, percentStarted <dbl>, percentOwned <dbl>,
#> #   percentChange <dbl>
#> 
#> $BOS
#> # A tibble: 16 × 17
#>    seasonId scoringPeriodId teamId abbrev lineupSlot actualSlot playerId
#>       <int>           <int>  <int> <fct>  <fct>      <fct>         <int>
#>  1     2023               1      2 BOS    QB         QB          3916387
#>  2     2023               1      2 BOS    RB         BE          4427366
#>  3     2023               1      2 BOS    RB         RB          3043078
#>  4     2023               1      2 BOS    WR         WR          3116406
#>  5     2023               1      2 BOS    WR         BE            15818
#>  6     2023               1      2 BOS    TE         TE          3116365
#>  7     2023               1      2 BOS    FLEX       BE          4047650
#>  8     2023               1      2 BOS    D/ST       D/ST         -16023
#>  9     2023               1      2 BOS    K          K           4360234
#> 10     2023               1      2 BOS    BE         BE          4045163
#> 11     2023               1      2 BOS    BE         FLEX        4241389
#> 12     2023               1      2 BOS    BE         BE          3116385
#> 13     2023               1      2 BOS    BE         BE          4379399
#> 14     2023               1      2 BOS    BE         RB          4047365
#> 15     2023               1      2 BOS    BE         WR          4362628
#> 16     2023               1      2 BOS    BE         BE          4248528
#> # ℹ 10 more variables: firstName <chr>, lastName <chr>, proTeam <fct>,
#> #   position <fct>, injuryStatus <chr>, projectedScore <dbl>,
#> #   actualScore <dbl>, percentStarted <dbl>, percentOwned <dbl>,
#> #   percentChange <dbl>
#> 
#> $CHI
#> # A tibble: 16 × 17
#>    seasonId scoringPeriodId teamId abbrev lineupSlot actualSlot playerId
#>       <int>           <int>  <int> <fct>  <fct>      <fct>         <int>
#>  1     2023               1      3 CHI    QB         BE          4038941
#>  2     2023               1      3 CHI    RB         BE          3042519
#>  3     2023               1      3 CHI    RB         RB          4239996
#>  4     2023               1      3 CHI    WR         WR          4262921
#>  5     2023               1      3 CHI    WR         FLEX        4374302
#>  6     2023               1      3 CHI    TE         TE            15847
#>  7     2023               1      3 CHI    FLEX       WR          4569618
#>  8     2023               1      3 CHI    D/ST       D/ST         -16025
#>  9     2023               1      3 CHI    K          K           3055899
#> 10     2023               1      3 CHI    BE         QB          4040715
#> 11     2023               1      3 CHI    BE         RB          3929630
#> 12     2023               1      3 CHI    BE         BE          4429795
#> 13     2023               1      3 CHI    BE         BE          4697815
#> 14     2023               1      3 CHI    BE         BE          2976499
#> 15     2023               1      3 CHI    BE         BE          3915511
#> 16     2023               1      3 CHI    BE         BE          3054850
#> # ℹ 10 more variables: firstName <chr>, lastName <chr>, proTeam <fct>,
#> #   position <fct>, injuryStatus <chr>, projectedScore <dbl>,
#> #   actualScore <dbl>, percentStarted <dbl>, percentOwned <dbl>,
#> #   percentChange <dbl>
#> 
#> $DEN
#> # A tibble: 16 × 17
#>    seasonId scoringPeriodId teamId abbrev lineupSlot actualSlot playerId
#>       <int>           <int>  <int> <fct>  <fct>      <fct>         <int>
#>  1     2023               1      4 DEN    QB         QB          3918298
#>  2     2023               1      4 DEN    RB         RB          3068267
#>  3     2023               1      4 DEN    RB         RB          3128720
#>  4     2023               1      4 DEN    WR         WR          2976212
#>  5     2023               1      4 DEN    WR         BE          3925357
#>  6     2023               1      4 DEN    TE         TE          4036133
#>  7     2023               1      4 DEN    FLEX       FLEX        4047646
#>  8     2023               1      4 DEN    D/ST       D/ST         -16002
#>  9     2023               1      4 DEN    K          K           3051909
#> 10     2023               1      4 DEN    BE         BE          4569173
#> 11     2023               1      4 DEN    BE         BE          4048244
#> 12     2023               1      4 DEN    BE         WR            16800
#> 13     2023               1      4 DEN    BE         BE          3045147
#> 14     2023               1      4 DEN    BE         BE          3932905
#> 15     2023               1      4 DEN    BE         BE          4241457
#> 16     2023               1      4 DEN    BE         BE          4239993
#> # ℹ 10 more variables: firstName <chr>, lastName <chr>, proTeam <fct>,
#> #   position <fct>, injuryStatus <chr>, projectedScore <dbl>,
#> #   actualScore <dbl>, percentStarted <dbl>, percentOwned <dbl>,
#> #   percentChange <dbl>
#>