Sort the optimal fantasy roster

Uses the roster settings for each league to find the best possible combinations of players to score the most fantasy points.

best_roster(
  leagueId = ffl_id(),
  useScore = c("actualScore", "projectedScore"),
  scoringPeriodId = NULL,
  ...
)

Arguments

leagueId
Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

useScore
One of "projectedScore" or "actualScore" (default).

scoringPeriodId
Integer week of NFL season. By default, NULL will use the current week (see ffl_week()). Scoring periods are always one week in length, whereas matchups might be longer.

...
Additional queries passed to httr::GET(). Arguments are converted to a named list and passed to query alongside view.

Value

A dataframe (or list) with optimal rosters.

Details

If scoringPeriodId is the current week (the default), then actual scoring might be incomplete (see projectedScore argument).

See also

Examples

best_roster(leagueId = "42654852", scoringPeriodId = 1)

#> $AUS
#> # A tibble: 16 × 17
#>    seasonId scoringPeriodId teamId abbrev lineupSlot actualSlot playerId
#>       <int>           <int>  <int> <fct>  <fct>      <fct>         <int>
#>  1     2021               1      1 AUS    QB         QB          3139477
#>  2     2021               1      1 AUS    RB         RB          3117251
#>  3     2021               1      1 AUS    RB         BE          4259545
#>  4     2021               1      1 AUS    WR         BE            16460
#>  5     2021               1      1 AUS    WR         BE          3932905
#>  6     2021               1      1 AUS    TE         TE          3116365
#>  7     2021               1      1 AUS    FLEX       RB          4242335
#>  8     2021               1      1 AUS    D/ST       D/ST         -16027
#>  9     2021               1      1 AUS    K          K           2473037
#> 10     2021               1      1 AUS    BE         BE          4045163
#> 11     2021               1      1 AUS    BE         BE          3919596
#> 12     2021               1      1 AUS    BE         FLEX        4360294
#> 13     2021               1      1 AUS    BE         WR            16800
#> 14     2021               1      1 AUS    BE         WR          3925357
#> 15     2021               1      1 AUS    BE         BE            16799
#> 16     2021               1      1 AUS    BE         BE          3128429
#> # … with 10 more variables: firstName <chr>, lastName <chr>, proTeam <fct>,
#> #   position <fct>, injuryStatus <chr>, projectedScore <dbl>,
#> #   actualScore <dbl>, percentStarted <dbl>, percentOwned <dbl>,
#> #   percentChange <dbl>
#> 
#> $BOS
#> # A tibble: 16 × 17
#>    seasonId scoringPeriodId teamId abbrev lineupSlot actualSlot playerId
#>       <int>           <int>  <int> <fct>  <fct>      <fct>         <int>
#>  1     2021               1      2 BOS    QB         QB          2577417
#>  2     2021               1      2 BOS    RB         FLEX        3128720
#>  3     2021               1      2 BOS    RB         RB          3116593
#>  4     2021               1      2 BOS    WR         BE          2976499
#>  5     2021               1      2 BOS    WR         WR          3116406
#>  6     2021               1      2 BOS    TE         TE          3040151
#>  7     2021               1      2 BOS    FLEX       BE          2977187
#>  8     2021               1      2 BOS    D/ST       D/ST         -16023
#>  9     2021               1      2 BOS    K          K             16339
#> 10     2021               1      2 BOS    BE         WR          4047650
#> 11     2021               1      2 BOS    BE         BE          4039359
#> 12     2021               1      2 BOS    BE         BE          2974858
#> 13     2021               1      2 BOS    BE         BE          4242214
#> 14     2021               1      2 BOS    BE         BE          3121422
#> 15     2021               1      2 BOS    BE         RB          3051392
#> 16     2021               1      2 BOS    BE         BE            13982
#> # … with 10 more variables: firstName <chr>, lastName <chr>, proTeam <fct>,
#> #   position <fct>, injuryStatus <chr>, projectedScore <dbl>,
#> #   actualScore <dbl>, percentStarted <dbl>, percentOwned <dbl>,
#> #   percentChange <dbl>
#> 
#> $CHI
#> # A tibble: 16 × 17
#>    seasonId scoringPeriodId teamId abbrev lineupSlot actualSlot playerId
#>       <int>           <int>  <int> <fct>  <fct>      <fct>         <int>
#>  1     2021               1      3 CHI    QB         QB          3918298
#>  2     2021               1      3 CHI    RB         BE          3116385
#>  3     2021               1      3 CHI    RB         BE          4047365
#>  4     2021               1      3 CHI    WR         WR            15795
#>  5     2021               1      3 CHI    WR         BE          4362628
#>  6     2021               1      3 CHI    TE         TE          2576925
#>  7     2021               1      3 CHI    FLEX       BE            15818
#>  8     2021               1      3 CHI    D/ST       D/ST         -16033
#>  9     2021               1      3 CHI    K          K             15683
#> 10     2021               1      3 CHI    BE         BE          3886818
#> 11     2021               1      3 CHI    BE         WR          4262921
#> 12     2021               1      3 CHI    BE         FLEX        3068267
#> 13     2021               1      3 CHI    BE         RB          3043078
#> 14     2021               1      3 CHI    BE         BE            16737
#> 15     2021               1      3 CHI    BE         RB          3929630
#> 16     2021               1      3 CHI    BE         BE          4360438
#> # … with 10 more variables: firstName <chr>, lastName <chr>, proTeam <fct>,
#> #   position <fct>, injuryStatus <chr>, projectedScore <dbl>,
#> #   actualScore <dbl>, percentStarted <dbl>, percentOwned <dbl>,
#> #   percentChange <dbl>
#> 
#> $DEN
#> # A tibble: 16 × 17
#>    seasonId scoringPeriodId teamId abbrev lineupSlot actualSlot playerId
#>       <int>           <int>  <int> <fct>  <fct>      <fct>         <int>
#>  1     2021               1      4 DEN    QB         QB          3917315
#>  2     2021               1      4 DEN    RB         BE          4035538
#>  3     2021               1      4 DEN    RB         RB          3054850
#>  4     2021               1      4 DEN    WR         BE          2577327
#>  5     2021               1      4 DEN    WR         BE          3116165
#>  6     2021               1      4 DEN    TE         TE            15847
#>  7     2021               1      4 DEN    FLEX       BE          4241389
#>  8     2021               1      4 DEN    D/ST       D/ST         -16025
#>  9     2021               1      4 DEN    K          K           3055899
#> 10     2021               1      4 DEN    BE         WR          2976212
#> 11     2021               1      4 DEN    BE         BE          3915416
#> 12     2021               1      4 DEN    BE         WR          4047646
#> 13     2021               1      4 DEN    BE         BE            15880
#> 14     2021               1      4 DEN    BE         FLEX        4241457
#> 15     2021               1      4 DEN    BE         RB          3042519
#> 16     2021               1      4 DEN    BE         BE          4241985
#> # … with 10 more variables: firstName <chr>, lastName <chr>, proTeam <fct>,
#> #   position <fct>, injuryStatus <chr>, projectedScore <dbl>,
#> #   actualScore <dbl>, percentStarted <dbl>, percentOwned <dbl>,
#> #   percentChange <dbl>
#>