Fantasy league teams

The emails, chats, notes, and messages sent by league members.

league_messages(leagueId = ffl_id(), leagueHistory = FALSE, ...)

Arguments

leagueId

Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

leagueHistory

logical; Should the leagueHistory version of the API be called? If TRUE, a list of results is returned, with one element for each historical year of the league.

...

Additional queries passed to httr::GET(). Arguments are converted to a named list and passed to query alongside view.

Value

A tibble of messages.

See also

Examples

league_messages(leagueId = "42654852")
#> # A tibble: 1 × 7
#>   id       type  author          date                content messages viewableBy
#>   <chr>    <chr> <chr>           <dttm>              <chr>   <list>   <list>    
#> 1 5af42ec9 NOTE  {22DFE7FF-9DF2… 2021-09-13 23:46:07 This [… <NULL>   <NULL>