Roster acquisition history

The date and method of each player's acquisition onto a fantasy roster.

player_acquire(leagueId = ffl_id(), leagueHistory = FALSE, ...)

Arguments

leagueId

Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

leagueHistory

logical; Should the leagueHistory version of the API be called? If TRUE, a list of results is returned, with one element for each historical year of the league.

...

Additional queries passed to httr::GET(). Arguments are converted to a named list and passed to query alongside view.

Value

A data frame of roster players with acquisition method and date.

Examples

player_acquire(leagueId = "42654852")
#> $AUS
#> # A tibble: 16 × 12
#>    seasonId scoringPeriodId teamId lineupSlot playerId firstName  lastName 
#>       <int>           <int>  <int> <fct>         <int> <chr>      <chr>    
#>  1     2021              19      1 RB          4242335 Jonathan   Taylor   
#>  2     2021              19      1 RB          4259545 D'Andre    Swift    
#>  3     2021              19      1 WR            16800 Davante    Adams    
#>  4     2021              19      1 WR            16460 Adam       Thielen  
#>  5     2021              19      1 TE          3116365 Mark       Andrews  
#>  6     2021              19      1 FLEX        3932905 Diontae    Johnson  
#>  7     2021              19      1 D/ST         -16027 Buccaneers D/ST     
#>  8     2021              19      1 K           2473037 Jason      Myers    
#>  9     2021              19      1 BE          3117251 Christian  McCaffrey
#> 10     2021              19      1 BE          4360294 Antonio    Gibson   
#> 11     2021              19      1 BE          3925357 Calvin     Ridley   
#> 12     2021              19      1 BE          4045163 Miles      Sanders  
#> 13     2021              19      1 BE          3919596 Chris      Carson   
#> 14     2021              19      1 BE          3128429 Courtland  Sutton   
#> 15     2021              19      1 BE          2577417 Dak        Prescott 
#> 16     2021              19      1 BE          3128720 Nick       Chubb    
#> # … with 5 more variables: proTeam <fct>, position <fct>,
#> #   acquisitionType <chr>, acquisitionDate <dttm>, onTeamId <fct>
#> 
#> $BOS
#> # A tibble: 15 × 12
#>    seasonId scoringPeriodId teamId lineupSlot playerId firstName lastName       
#>       <int>           <int>  <int> <fct>         <int> <chr>     <chr>          
#>  1     2021              19      2 RB          3116593 Dalvin    Cook           
#>  2     2021              19      2 WR          4047650 DK        Metcalf        
#>  3     2021              19      2 WR          2977187 Cooper    Kupp           
#>  4     2021              19      2 FLEX        3116406 Tyreek    Hill           
#>  5     2021              19      2 D/ST         -16023 Steelers  D/ST           
#>  6     2021              19      2 K             16339 Brandon   McManus        
#>  7     2021              19      2 BE          3051392 Ezekiel   Elliott        
#>  8     2021              19      2 BE          4242214 Clyde     Edwards-Helaire
#>  9     2021              19      2 BE          3121422 Terry     McLaurin       
#> 10     2021              19      2 BE          3040151 George    Kittle         
#> 11     2021              19      2 BE          2976499 Amari     Cooper         
#> 12     2021              19      2 BE          4039359 Darrell   Henderson Jr.  
#> 13     2021              19      2 BE            13982 Julio     Jones          
#> 14     2021              19      2 BE          2974858 Kenny     Golladay       
#> 15     2021              19      2 BE          3139477 Patrick   Mahomes        
#> # … with 5 more variables: proTeam <fct>, position <fct>,
#> #   acquisitionType <chr>, acquisitionDate <dttm>, onTeamId <fct>
#> 
#> $CHI
#> # A tibble: 16 × 12
#>    seasonId scoringPeriodId teamId lineupSlot playerId firstName lastName 
#>       <int>           <int>  <int> <fct>         <int> <chr>     <chr>    
#>  1     2021              19      3 QB          3918298 Josh      Allen    
#>  2     2021              19      3 RB          3043078 Derrick   Henry    
#>  3     2021              19      3 RB          3116385 Joe       Mixon    
#>  4     2021              19      3 WR          4262921 Justin    Jefferson
#>  5     2021              19      3 WR            15818 Keenan    Allen    
#>  6     2021              19      3 TE          2576925 Darren    Waller   
#>  7     2021              19      3 FLEX        3068267 Austin    Ekeler   
#>  8     2021              19      3 D/ST         -16002 Bills     D/ST     
#>  9     2021              19      3 K             15683 Justin    Tucker   
#> 10     2021              19      3 BE          3929630 Saquon    Barkley  
#> 11     2021              19      3 BE            15795 DeAndre   Hopkins  
#> 12     2021              19      3 BE            16737 Mike      Evans    
#> 13     2021              19      3 BE          4047365 Josh      Jacobs   
#> 14     2021              19      3 BE          3886818 Myles     Gaskin   
#> 15     2021              19      3 BE          4362628 Ja'Marr   Chase    
#> 16     2021              19      3 BE          4360438 Brandon   Aiyuk    
#> # … with 5 more variables: proTeam <fct>, position <fct>,
#> #   acquisitionType <chr>, acquisitionDate <dttm>, onTeamId <fct>
#> 
#> $DEN
#> # A tibble: 17 × 12
#>    seasonId scoringPeriodId teamId lineupSlot playerId firstName lastName  
#>       <int>           <int>  <int> <fct>         <int> <chr>     <chr>     
#>  1     2021              19      4 QB          3917315 Kyler     Murray    
#>  2     2021              19      4 RB          3042519 Aaron     Jones     
#>  3     2021              19      4 RB          2979843 Kenyan    Drake     
#>  4     2021              19      4 WR          2976212 Stefon    Diggs     
#>  5     2021              19      4 WR          2577327 Tyler     Lockett   
#>  6     2021              19      4 TE            15847 Travis    Kelce     
#>  7     2021              19      4 FLEX        4241457 Najee     Harris    
#>  8     2021              19      4 D/ST         -16007 Broncos   D/ST      
#>  9     2021              19      4 K           3055899 Harrison  Butker    
#> 10     2021              19      4 BE          3054850 Alvin     Kamara    
#> 11     2021              19      4 BE          4047646 A.J.      Brown     
#> 12     2021              19      4 BE          4035538 David     Montgomery
#> 13     2021              19      4 BE            15880 Robert    Woods     
#> 14     2021              19      4 BE          4241389 CeeDee    Lamb      
#> 15     2021              19      4 BE          3915416 DJ        Moore     
#> 16     2021              19      4 BE          3116165 Chris     Godwin    
#> 17     2021              19      4 IR          4241985 J.K.      Dobbins   
#> # … with 5 more variables: proTeam <fct>, position <fct>,
#> #   acquisitionType <chr>, acquisitionDate <dttm>, onTeamId <fct>
#>