Roster acquisition history

The date and method of each player's acquisition onto a fantasy roster.

player_acquire(leagueId = ffl_id(), leagueHistory = FALSE, ...)

Arguments

leagueId

Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

leagueHistory

logical; Should the leagueHistory version of the API be called? If TRUE, a list of results is returned, with one element for each historical year of the league.

...

Additional queries passed to httr::GET(). Arguments are converted to a named list and passed to query alongside view.

Value

A data frame of roster players with acquisition method and date.

Examples

player_acquire(leagueId = "42654852")
#> $AUS
#> # A tibble: 16 × 12
#>    seasonId scoringPeriodId teamId lineupSlot playerId firstName lastName  
#>       <int>           <int>  <int> <fct>         <int> <chr>     <chr>     
#>  1     2023               2      1 QB          3139477 Patrick   Mahomes   
#>  2     2023               2      1 RB          3117251 Christian McCaffrey 
#>  3     2023               2      1 RB          4430807 Bijan     Robinson  
#>  4     2023               2      1 WR          2977187 Cooper    Kupp      
#>  5     2023               2      1 WR          4372016 Jaylen    Waddle    
#>  6     2023               2      1 TE          2576925 Darren    Waller    
#>  7     2023               2      1 FLEX        3916148 Tony      Pollard   
#>  8     2023               2      1 D/ST         -16018 Saints    D/ST      
#>  9     2023               2      1 K             15683 Justin    Tucker    
#> 10     2023               2      1 BE          4361370 Chris     Olave     
#> 11     2023               2      1 BE          4241478 DeVonta   Smith     
#> 12     2023               2      1 BE          4360238 Dameon    Pierce    
#> 13     2023               2      1 BE          3126486 Deebo     Samuel    
#> 14     2023               2      1 BE            15795 DeAndre   Hopkins   
#> 15     2023               2      1 BE          3116165 Chris     Godwin    
#> 16     2023               2      1 BE          4567048 Kenneth   Walker III
#> # ℹ 5 more variables: proTeam <fct>, position <fct>, acquisitionType <chr>,
#> #   acquisitionDate <dttm>, onTeamId <fct>
#> 
#> $BOS
#> # A tibble: 16 × 12
#>    seasonId scoringPeriodId teamId lineupSlot playerId firstName lastName 
#>       <int>           <int>  <int> <fct>         <int> <chr>     <chr>    
#>  1     2023               2      2 QB          3916387 Lamar     Jackson  
#>  2     2023               2      2 RB          3043078 Derrick   Henry    
#>  3     2023               2      2 RB          4047365 Josh      Jacobs   
#>  4     2023               2      2 WR          4362628 Ja'Marr   Chase    
#>  5     2023               2      2 WR          3116406 Tyreek    Hill     
#>  6     2023               2      2 TE          3116365 Mark      Andrews  
#>  7     2023               2      2 FLEX        4241389 CeeDee    Lamb     
#>  8     2023               2      2 D/ST         -16023 Steelers  D/ST     
#>  9     2023               2      2 K           4360234 Evan      McPherson
#> 10     2023               2      2 BE          3116385 Joe       Mixon    
#> 11     2023               2      2 BE          4047650 DK        Metcalf  
#> 12     2023               2      2 BE            15818 Keenan    Allen    
#> 13     2023               2      2 BE          4248528 Christian Watson   
#> 14     2023               2      2 BE          4427366 Breece    Hall     
#> 15     2023               2      2 BE          4045163 Miles     Sanders  
#> 16     2023               2      2 BE          4379399 James     Cook     
#> # ℹ 5 more variables: proTeam <fct>, position <fct>, acquisitionType <chr>,
#> #   acquisitionDate <dttm>, onTeamId <fct>
#> 
#> $CHI
#> # A tibble: 16 × 12
#>    seasonId scoringPeriodId teamId lineupSlot playerId firstName lastName   
#>       <int>           <int>  <int> <fct>         <int> <chr>     <chr>      
#>  1     2023               2      3 QB          4040715 Jalen     Hurts      
#>  2     2023               2      3 RB          3929630 Saquon    Barkley    
#>  3     2023               2      3 RB          4239996 Travis    Etienne Jr.
#>  4     2023               2      3 WR          4262921 Justin    Jefferson  
#>  5     2023               2      3 WR          4569618 Garrett   Wilson     
#>  6     2023               2      3 TE            15847 Travis    Kelce      
#>  7     2023               2      3 FLEX        4374302 Amon-Ra   St. Brown  
#>  8     2023               2      3 D/ST         -16025 49ers     D/ST       
#>  9     2023               2      3 K           3055899 Harrison  Butker     
#> 10     2023               2      3 BE          4429795 Jahmyr    Gibbs      
#> 11     2023               2      3 BE          3042519 Aaron     Jones      
#> 12     2023               2      3 BE          3915511 Joe       Burrow     
#> 13     2023               2      3 BE          2976499 Amari     Cooper     
#> 14     2023               2      3 BE          4697815 Rachaad   White      
#> 15     2023               2      3 BE          3054850 Alvin     Kamara     
#> 16     2023               2      3 BE          4038941 Justin    Herbert    
#> # ℹ 5 more variables: proTeam <fct>, position <fct>, acquisitionType <chr>,
#> #   acquisitionDate <dttm>, onTeamId <fct>
#> 
#> $DEN
#> # A tibble: 16 × 12
#>    seasonId scoringPeriodId teamId lineupSlot playerId firstName lastName 
#>       <int>           <int>  <int> <fct>         <int> <chr>     <chr>    
#>  1     2023               2      4 QB          3918298 Josh      Allen    
#>  2     2023               2      4 RB          3068267 Austin    Ekeler   
#>  3     2023               2      4 RB          3128720 Nick      Chubb    
#>  4     2023               2      4 WR            16800 Davante   Adams    
#>  5     2023               2      4 WR          2976212 Stefon    Diggs    
#>  6     2023               2      4 TE          4036133 T.J.      Hockenson
#>  7     2023               2      4 FLEX        4047646 A.J.      Brown    
#>  8     2023               2      4 D/ST         -16002 Bills     D/ST     
#>  9     2023               2      4 K           3051909 Daniel    Carlson  
#> 10     2023               2      4 BE          4239993 Tee       Higgins  
#> 11     2023               2      4 BE          4241457 Najee     Harris   
#> 12     2023               2      4 BE          4569173 Rhamondre Stevenson
#> 13     2023               2      4 BE          3925357 Calvin    Ridley   
#> 14     2023               2      4 BE          3932905 Diontae   Johnson  
#> 15     2023               2      4 BE          3045147 James     Conner   
#> 16     2023               2      4 BE          4048244 Alexander Mattison 
#> # ℹ 5 more variables: proTeam <fct>, position <fct>, acquisitionType <chr>,
#> #   acquisitionDate <dttm>, onTeamId <fct>
#>