Roster acquisition history

The date and method of each player's acquisition onto a fantasy roster.

player_acquire(leagueId = ffl_id(), leagueHistory = FALSE, ...)

Arguments

leagueId
Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

leagueHistory
logical; Should the leagueHistory version of the API be called? If TRUE, a list of results is returned, with one element for each historical year of the league.

...
Additional queries passed to httr::GET() (e.g., scoringPeriodId). Arguments are converted to a named list to be passed alongside view.

Value

A data frame of roster players with acquisition method and date.

Examples

player_acquire(leagueId = "42654852")

#> $AUS
#> # A tibble: 16 × 11
#>    seasonId scoringPeriodId onTeamId lineupSlotId      id firstName  lastName   
#>       <int>           <int> <fct>    <fct>          <int> <chr>      <chr>      
#>  1     2021               5 AUS      QB           3139477 Patrick    Mahomes    
#>  2     2021               5 AUS      RB           4242335 Jonathan   Taylor     
#>  3     2021               5 AUS      RB           4259545 D'Andre    Swift      
#>  4     2021               5 AUS      WR             16800 Davante    Adams      
#>  5     2021               5 AUS      WR           3925357 Calvin     Ridley     
#>  6     2021               5 AUS      TE           3116365 Mark       Andrews    
#>  7     2021               5 AUS      FX           3932905 Diontae    Johnson    
#>  8     2021               5 AUS      DS            -16027 Buccaneers D/ST       
#>  9     2021               5 AUS      PK           2473037 Jason      Myers      
#> 10     2021               5 AUS      BE           3117251 Christian  McCaffrey  
#> 11     2021               5 AUS      BE           4360294 Antonio    Gibson     
#> 12     2021               5 AUS      BE           4045163 Miles      Sanders    
#> 13     2021               5 AUS      BE           3919596 Chris      Carson     
#> 14     2021               5 AUS      BE             16799 Allen      Robinson II
#> 15     2021               5 AUS      BE             16460 Adam       Thielen    
#> 16     2021               5 AUS      BE           3128429 Courtland  Sutton     
#> # … with 4 more variables: proTeamId <fct>, defaultPositionId <fct>,
#> #   acquisitionType <chr>, acquisitionDate <dttm>
#> 
#> $BOS
#> # A tibble: 16 × 11
#>    seasonId scoringPeriodId onTeamId lineupSlotId      id firstName lastName    
#>       <int>           <int> <fct>    <fct>          <int> <chr>     <chr>       
#>  1     2021               5 BOS      QB           2577417 Dak       Prescott    
#>  2     2021               5 BOS      RB           3116593 Dalvin    Cook        
#>  3     2021               5 BOS      RB           3128720 Nick      Chubb       
#>  4     2021               5 BOS      WR           4047650 DK        Metcalf     
#>  5     2021               5 BOS      WR           2977187 Cooper    Kupp        
#>  6     2021               5 BOS      TE           3040151 George    Kittle      
#>  7     2021               5 BOS      FX           3116406 Tyreek    Hill        
#>  8     2021               5 BOS      DS            -16023 Steelers  D/ST        
#>  9     2021               5 BOS      PK             16339 Brandon   McManus     
#> 10     2021               5 BOS      BE           3051392 Ezekiel   Elliott     
#> 11     2021               5 BOS      BE           4242214 Clyde     Edwards-Hel…
#> 12     2021               5 BOS      BE           3121422 Terry     McLaurin    
#> 13     2021               5 BOS      BE           2976499 Amari     Cooper      
#> 14     2021               5 BOS      BE           4039359 Darrell   Henderson J…
#> 15     2021               5 BOS      BE             13982 Julio     Jones       
#> 16     2021               5 BOS      BE           2974858 Kenny     Golladay    
#> # … with 4 more variables: proTeamId <fct>, defaultPositionId <fct>,
#> #   acquisitionType <chr>, acquisitionDate <dttm>
#> 
#> $CHI
#> # A tibble: 16 × 11
#>    seasonId scoringPeriodId onTeamId lineupSlotId      id firstName lastName 
#>       <int>           <int> <fct>    <fct>          <int> <chr>     <chr>    
#>  1     2021               5 CHI      QB           3918298 Josh      Allen    
#>  2     2021               5 CHI      RB           3043078 Derrick   Henry    
#>  3     2021               5 CHI      RB           3116385 Joe       Mixon    
#>  4     2021               5 CHI      WR           4262921 Justin    Jefferson
#>  5     2021               5 CHI      WR             15818 Keenan    Allen    
#>  6     2021               5 CHI      TE           2576925 Darren    Waller   
#>  7     2021               5 CHI      FX           3068267 Austin    Ekeler   
#>  8     2021               5 CHI      PK             15683 Justin    Tucker   
#>  9     2021               5 CHI      BE           3929630 Saquon    Barkley  
#> 10     2021               5 CHI      BE             15795 DeAndre   Hopkins  
#> 11     2021               5 CHI      BE             16737 Mike      Evans    
#> 12     2021               5 CHI      BE           4047365 Josh      Jacobs   
#> 13     2021               5 CHI      BE           3886818 Myles     Gaskin   
#> 14     2021               5 CHI      BE           4362628 Ja'Marr   Chase    
#> 15     2021               5 CHI      BE           4360438 Brandon   Aiyuk    
#> 16     2021               5 CHI      BE            -16002 Bills     D/ST     
#> # … with 4 more variables: proTeamId <fct>, defaultPositionId <fct>,
#> #   acquisitionType <chr>, acquisitionDate <dttm>
#> 
#> $DEN
#> # A tibble: 17 × 11
#>    seasonId scoringPeriodId onTeamId lineupSlotId      id firstName lastName  
#>       <int>           <int> <fct>    <fct>          <int> <chr>     <chr>     
#>  1     2021               5 DEN      QB           3917315 Kyler     Murray    
#>  2     2021               5 DEN      RB           3054850 Alvin     Kamara    
#>  3     2021               5 DEN      RB           3042519 Aaron     Jones     
#>  4     2021               5 DEN      WR           2976212 Stefon    Diggs     
#>  5     2021               5 DEN      WR           2577327 Tyler     Lockett   
#>  6     2021               5 DEN      TE             15847 Travis    Kelce     
#>  7     2021               5 DEN      FX           4241457 Najee     Harris    
#>  8     2021               5 DEN      PK           3055899 Harrison  Butker    
#>  9     2021               5 DEN      BE           4047646 A.J.      Brown     
#> 10     2021               5 DEN      BE           4035538 David     Montgomery
#> 11     2021               5 DEN      BE             15880 Robert    Woods     
#> 12     2021               5 DEN      BE           4241389 CeeDee    Lamb      
#> 13     2021               5 DEN      BE           3915416 DJ        Moore     
#> 14     2021               5 DEN      BE           3116165 Chris     Godwin    
#> 15     2021               5 DEN      BE           2979843 Kenyan    Drake     
#> 16     2021               5 DEN      BE            -16007 Broncos   D/ST      
#> 17     2021               5 DEN      IR           4241985 J.K.      Dobbins   
#> # … with 4 more variables: proTeamId <fct>, defaultPositionId <fct>,
#> #   acquisitionType <chr>, acquisitionDate <dttm>
#>