Player outlooks

All available weekly ESPN outlook writeups for NFL players.

player_outlook(leagueId = ffl_id(), limit = 50)

Arguments

leagueId

Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

limit

The limit of players to return. Use "" or NULL to return all. Defaults to 50, which is the default limit used by ESPN. Removing the limit can make the request take a long time.

Value

A data frame of player outlooks by scoring period.

See also

Examples

player_outlook()
#> # A tibble: 144 × 6
#>    seasonId scoringPeriodId      id firstName lastName outlook                  
#>       <int>           <int>   <int> <chr>     <chr>    <chr>                    
#>  1     2022               0 2977187 Cooper    Kupp     "Here is an incomplete l…
#>  2     2022               1 2977187 Cooper    Kupp     "Kupp was \"other-wordly…
#>  3     2022               2 2977187 Cooper    Kupp     "Starting to think this …
#>  4     2022               0 4242335 Jonathan  Taylor   "Taylor returns as the f…
#>  5     2022               1 4242335 Jonathan  Taylor   "Taylor is coming off a …
#>  6     2022               2 4242335 Jonathan  Taylor   "Taylor posted the secon…
#>  7     2022               0 2976212 Stefon    Diggs    "Diggs followed up a car…
#>  8     2022               1 2976212 Stefon    Diggs    "Diggs made 103 receptio…
#>  9     2022               2 2976212 Stefon    Diggs    "The superstar wideout l…
#> 10     2022               0 3918298 Josh      Allen    "Allen enters his fifth …
#> # … with 134 more rows