Fantasy team rosters

The roster of all teams in a league.

team_roster(
  leagueId = ffl_id(),
  leagueHistory = FALSE,
  scoringPeriodId = NULL,
  ...
)

Arguments

leagueId

Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

leagueHistory

logical; Should the leagueHistory version of the API be called? If TRUE, a list of results is returned, with one element for each historical year of the league.

scoringPeriodId

Integer week of NFL season. By default, NULL will use the current week (see ffl_week()). Scoring periods are always one week in length, whereas matchups might be longer.

...

Additional queries passed to httr::GET(). Arguments are converted to a named list and passed to query alongside view.

Value

A dataframe (or list) with league teams.

See also

Examples

team_roster(leagueId = "42654852", scoringPeriodId = 1)
#> $AUS
#> # A tibble: 16 × 16
#>    seaso…¹ scori…² teamId abbrev lineu…³ playe…⁴ first…⁵ lastN…⁶ proTeam posit…⁷
#>      <int>   <int>  <int> <fct>  <fct>     <int> <chr>   <chr>   <fct>   <fct>  
#>  1    2022       1      1 AUS    QB      3916387 Lamar   Jackson Bal     QB     
#>  2    2022       1      1 AUS    RB      3117251 Christ… McCaff… Car     RB     
#>  3    2022       1      1 AUS    RB      3054850 Alvin   Kamara  NO      RB     
#>  4    2022       1      1 AUS    WR      4362628 Ja'Marr Chase   Cin     WR     
#>  5    2022       1      1 AUS    WR        15818 Keenan  Allen   LAC     WR     
#>  6    2022       1      1 AUS    TE      3116365 Mark    Andrews Bal     TE     
#>  7    2022       1      1 AUS    FLEX    4259545 D'Andre Swift   Det     RB     
#>  8    2022       1      1 AUS    D/ST     -16009 Packers D/ST    GB      D/ST   
#>  9    2022       1      1 AUS    K       4249087 Matt    Gay     LAR     K      
#> 10    2022       1      1 AUS    BE        16737 Mike    Evans   TB      WR     
#> 11    2022       1      1 AUS    BE      4047646 A.J.    Brown   Phi     WR     
#> 12    2022       1      1 AUS    BE      3128720 Nick    Chubb   Cle     RB     
#> 13    2022       1      1 AUS    BE        16731 Brandin Cooks   Hou     WR     
#> 14    2022       1      1 AUS    BE      4239996 Travis  Etienn… Jax     RB     
#> 15    2022       1      1 AUS    BE      4241463 Jerry   Jeudy   Den     WR     
#> 16    2022       1      1 AUS    BE      4241985 J.K.    Dobbins Bal     RB     
#> # … with 6 more variables: injuryStatus <chr>, projectedScore <dbl>,
#> #   actualScore <dbl>, percentStarted <dbl>, percentOwned <dbl>,
#> #   percentChange <dbl>, and abbreviated variable names ¹​seasonId,
#> #   ²​scoringPeriodId, ³​lineupSlot, ⁴​playerId, ⁵​firstName, ⁶​lastName, ⁷​position
#> 
#> $BOS
#> # A tibble: 16 × 16
#>    seaso…¹ scori…² teamId abbrev lineu…³ playe…⁴ first…⁵ lastN…⁶ proTeam posit…⁷
#>      <int>   <int>  <int> <fct>  <fct>     <int> <chr>   <chr>   <fct>   <fct>  
#>  1    2022       1      2 BOS    QB      3918298 Josh    Allen   Buf     QB     
#>  2    2022       1      2 BOS    RB      3068267 Austin  Ekeler  LAC     RB     
#>  3    2022       1      2 BOS    RB      4241457 Najee   Harris  Pit     RB     
#>  4    2022       1      2 BOS    WR      3126486 Deebo   Samuel  SF      WR     
#>  5    2022       1      2 BOS    WR      2976212 Stefon  Diggs   Buf     WR     
#>  6    2022       1      2 BOS    TE        15847 Travis  Kelce   KC      TE     
#>  7    2022       1      2 BOS    FLEX    4361579 Javonte Willia… Den     RB     
#>  8    2022       1      2 BOS    D/ST     -16027 Buccan… D/ST    TB      D/ST   
#>  9    2022       1      2 BOS    K         15683 Justin  Tucker  Bal     K      
#> 10    2022       1      2 BOS    BE      3929630 Saquon  Barkley NYG     RB     
#> 11    2022       1      2 BOS    BE      4240021 Cam     Akers   LAR     RB     
#> 12    2022       1      2 BOS    BE      4372016 Jaylen  Waddle  Mia     WR     
#> 13    2022       1      2 BOS    BE      4241372 Marqui… Brown   Ari     WR     
#> 14    2022       1      2 BOS    BE      4047650 DK      Metcalf Sea     WR     
#> 15    2022       1      2 BOS    BE      4374302 Amon-Ra St. Br… Det     WR     
#> 16    2022       1      2 BOS    BE      4040655 Darnell Mooney  Chi     WR     
#> # … with 6 more variables: injuryStatus <chr>, projectedScore <dbl>,
#> #   actualScore <dbl>, percentStarted <dbl>, percentOwned <dbl>,
#> #   percentChange <dbl>, and abbreviated variable names ¹​seasonId,
#> #   ²​scoringPeriodId, ³​lineupSlot, ⁴​playerId, ⁵​firstName, ⁶​lastName, ⁷​position
#> 
#> $CHI
#> # A tibble: 16 × 16
#>    seaso…¹ scori…² teamId abbrev lineu…³ playe…⁴ first…⁵ lastN…⁶ proTeam posit…⁷
#>      <int>   <int>  <int> <fct>  <fct>     <int> <chr>   <chr>   <fct>   <fct>  
#>  1    2022       1      3 CHI    QB      4038941 Justin  Herbert LAC     QB     
#>  2    2022       1      3 CHI    RB      4242335 Jonath… Taylor  Ind     RB     
#>  3    2022       1      3 CHI    RB      3116593 Dalvin  Cook    Min     RB     
#>  4    2022       1      3 CHI    WR      4262921 Justin  Jeffer… Min     WR     
#>  5    2022       1      3 CHI    WR      4241389 CeeDee  Lamb    Dal     WR     
#>  6    2022       1      3 CHI    TE      4360248 Kyle    Pitts   Atl     TE     
#>  7    2022       1      3 CHI    FLEX    3116406 Tyreek  Hill    Mia     WR     
#>  8    2022       1      3 CHI    D/ST     -16002 Bills   D/ST    Buf     D/ST   
#>  9    2022       1      3 CHI    K       3055899 Harris… Butker  KC      K      
#> 10    2022       1      3 CHI    BE      3042519 Aaron   Jones   GB      RB     
#> 11    2022       1      3 CHI    BE      4239993 Tee     Higgins Cin     WR     
#> 12    2022       1      3 CHI    BE      3121422 Terry   McLaur… Wsh     WR     
#> 13    2022       1      3 CHI    BE      4035538 David   Montgo… Chi     RB     
#> 14    2022       1      3 CHI    BE      3051392 Ezekiel Elliott Dal     RB     
#> 15    2022       1      3 CHI    BE      3128429 Courtl… Sutton  Den     WR     
#> 16    2022       1      3 CHI    BE      4243537 Gabe    Davis   Buf     WR     
#> # … with 6 more variables: injuryStatus <chr>, projectedScore <dbl>,
#> #   actualScore <dbl>, percentStarted <dbl>, percentOwned <dbl>,
#> #   percentChange <dbl>, and abbreviated variable names ¹​seasonId,
#> #   ²​scoringPeriodId, ³​lineupSlot, ⁴​playerId, ⁵​firstName, ⁶​lastName, ⁷​position
#> 
#> $DEN
#> # A tibble: 16 × 16
#>    seaso…¹ scori…² teamId abbrev lineu…³ playe…⁴ first…⁵ lastN…⁶ proTeam posit…⁷
#>      <int>   <int>  <int> <fct>  <fct>     <int> <chr>   <chr>   <fct>   <fct>  
#>  1    2022       1      4 DEN    QB      3139477 Patrick Mahomes KC      QB     
#>  2    2022       1      4 DEN    RB      3043078 Derrick Henry   Ten     RB     
#>  3    2022       1      4 DEN    RB      3116385 Joe     Mixon   Cin     RB     
#>  4    2022       1      4 DEN    WR      2977187 Cooper  Kupp    LAR     WR     
#>  5    2022       1      4 DEN    WR        16800 Davante Adams   LV      WR     
#>  6    2022       1      4 DEN    TE      2576925 Darren  Waller  LV      TE     
#>  7    2022       1      4 DEN    FLEX    3115364 Leonard Fourne… TB      RB     
#>  8    2022       1      4 DEN    D/ST     -16018 Saints  D/ST    NO      D/ST   
#>  9    2022       1      4 DEN    K       4360234 Evan    McPher… Cin     K      
#> 10    2022       1      4 DEN    BE      3045147 James   Conner  Ari     RB     
#> 11    2022       1      4 DEN    BE      3915416 DJ      Moore   Car     WR     
#> 12    2022       1      4 DEN    BE      4035687 Michael Pittma… Ind     WR     
#> 13    2022       1      4 DEN    BE      3932905 Diontae Johnson Pit     WR     
#> 14    2022       1      4 DEN    BE      3045138 Mike    Willia… LAC     WR     
#> 15    2022       1      4 DEN    BE      4427366 Breece  Hall    NYJ     RB     
#> 16    2022       1      4 DEN    BE      3116165 Chris   Godwin  TB      WR     
#> # … with 6 more variables: injuryStatus <chr>, projectedScore <dbl>,
#> #   actualScore <dbl>, percentStarted <dbl>, percentOwned <dbl>,
#> #   percentChange <dbl>, and abbreviated variable names ¹​seasonId,
#> #   ²​scoringPeriodId, ³​lineupSlot, ⁴​playerId, ⁵​firstName, ⁶​lastName, ⁷​position
#>