Fantasy team rosters

The roster of all teams in a league.

team_roster(leagueId = ffl_id(), leagueHistory = FALSE, ...)

Arguments

leagueId
Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

leagueHistory
logical; Should the leagueHistory version of the API be called? If TRUE, a list of results is returned, with one element for each historical year of the league.

...
Additional queries passed to httr::GET() (e.g., scoringPeriodId). Arguments are converted to a named list to be passed alongside view.

Value

A dataframe (or list) with league teams.

Examples

team_roster(leagueId = "42654852", scoringPeriodId = 1)

#> [[1]]
#> # A tibble: 16 × 15
#>    seasonId scoringPeriodId teamId lineupSlot      id firstName lastName proTeam
#>       <int>           <int>  <int> <fct>        <int> <chr>     <chr>    <fct>  
#>  1     2021               1      1 QB         3139477 Patrick   Mahomes  KC     
#>  2     2021               1      1 RB         3117251 Christian McCaffr… Car    
#>  3     2021               1      1 RB         4242335 Jonathan  Taylor   Ind    
#>  4     2021               1      1 WR           16800 Davante   Adams    GB     
#>  5     2021               1      1 WR         3925357 Calvin    Ridley   Atl    
#>  6     2021               1      1 TE         3116365 Mark      Andrews  Bal    
#>  7     2021               1      1 FLEX       4360294 Antonio   Gibson   Wsh    
#>  8     2021               1      1 D/ST        -16027 Buccanee… D/ST     TB     
#>  9     2021               1      1 K          2473037 Jason     Myers    Sea    
#> 10     2021               1      1 BE         4259545 D'Andre   Swift    Det    
#> 11     2021               1      1 BE         4045163 Miles     Sanders  Phi    
#> 12     2021               1      1 BE         3919596 Chris     Carson   Sea    
#> 13     2021               1      1 BE           16799 Allen     Robinso… Chi    
#> 14     2021               1      1 BE           16460 Adam      Thielen  Min    
#> 15     2021               1      1 BE         3932905 Diontae   Johnson  Pit    
#> 16     2021               1      1 BE         3128429 Courtland Sutton   Den    
#> # … with 7 more variables: position <fct>, injuryStatus <chr>,
#> #   projectedScore <dbl>, actualScore <dbl>, percentStarted <dbl>,
#> #   percentOwned <dbl>, percentChange <dbl>
#> 
#> [[2]]
#> # A tibble: 16 × 15
#>    seasonId scoringPeriodId teamId lineupSlot      id firstName lastName proTeam
#>       <int>           <int>  <int> <fct>        <int> <chr>     <chr>    <fct>  
#>  1     2021               1      2 QB         2577417 Dak       Prescott Dal    
#>  2     2021               1      2 RB         3116593 Dalvin    Cook     Min    
#>  3     2021               1      2 RB         3051392 Ezekiel   Elliott  Dal    
#>  4     2021               1      2 WR         3116406 Tyreek    Hill     KC     
#>  5     2021               1      2 WR         4047650 DK        Metcalf  Sea    
#>  6     2021               1      2 TE         3040151 George    Kittle   SF     
#>  7     2021               1      2 FLEX       3128720 Nick      Chubb    Cle    
#>  8     2021               1      2 D/ST        -16023 Steelers  D/ST     Pit    
#>  9     2021               1      2 K            16339 Brandon   McManus  Den    
#> 10     2021               1      2 BE         4242214 Clyde     Edwards… KC     
#> 11     2021               1      2 BE         3121422 Terry     McLaurin Wsh    
#> 12     2021               1      2 BE         2976499 Amari     Cooper   Dal    
#> 13     2021               1      2 BE         4039359 Darrell   Henders… LAR    
#> 14     2021               1      2 BE           13982 Julio     Jones    Ten    
#> 15     2021               1      2 BE         2977187 Cooper    Kupp     LAR    
#> 16     2021               1      2 BE         2974858 Kenny     Golladay NYG    
#> # … with 7 more variables: position <fct>, injuryStatus <chr>,
#> #   projectedScore <dbl>, actualScore <dbl>, percentStarted <dbl>,
#> #   percentOwned <dbl>, percentChange <dbl>
#> 
#> [[3]]
#> # A tibble: 16 × 15
#>    seasonId scoringPeriodId teamId lineupSlot      id firstName lastName proTeam
#>       <int>           <int>  <int> <fct>        <int> <chr>     <chr>    <fct>  
#>  1     2021               1      3 QB         3918298 Josh      Allen    Buf    
#>  2     2021               1      3 RB         3929630 Saquon    Barkley  NYG    
#>  3     2021               1      3 RB         3043078 Derrick   Henry    Ten    
#>  4     2021               1      3 WR           15795 DeAndre   Hopkins  Ari    
#>  5     2021               1      3 WR         4262921 Justin    Jeffers… Min    
#>  6     2021               1      3 TE         2576925 Darren    Waller   LV     
#>  7     2021               1      3 FLEX       3068267 Austin    Ekeler   LAC    
#>  8     2021               1      0 D/ST        -16033 Ravens    D/ST     Bal    
#>  9     2021               1      3 K            15683 Justin    Tucker   Bal    
#> 10     2021               1      3 BE         3116385 Joe       Mixon    Cin    
#> 11     2021               1      3 BE           15818 Keenan    Allen    LAC    
#> 12     2021               1      3 BE           16737 Mike      Evans    TB     
#> 13     2021               1      3 BE         4047365 Josh      Jacobs   LV     
#> 14     2021               1      3 BE         3886818 Myles     Gaskin   Mia    
#> 15     2021               1      3 BE         4362628 Ja'Marr   Chase    Cin    
#> 16     2021               1      3 BE         4360438 Brandon   Aiyuk    SF     
#> # … with 7 more variables: position <fct>, injuryStatus <chr>,
#> #   projectedScore <dbl>, actualScore <dbl>, percentStarted <dbl>,
#> #   percentOwned <dbl>, percentChange <dbl>
#> 
#> [[4]]
#> # A tibble: 16 × 15
#>    seasonId scoringPeriodId teamId lineupSlot      id firstName lastName proTeam
#>       <int>           <int>  <int> <fct>        <int> <chr>     <chr>    <fct>  
#>  1     2021               1      4 QB         3917315 Kyler     Murray   Ari    
#>  2     2021               1      4 RB         3054850 Alvin     Kamara   NO     
#>  3     2021               1      4 RB         3042519 Aaron     Jones    GB     
#>  4     2021               1      4 WR         2976212 Stefon    Diggs    Buf    
#>  5     2021               1      4 WR         4047646 A.J.      Brown    Ten    
#>  6     2021               1      4 TE           15847 Travis    Kelce    KC     
#>  7     2021               1      4 FLEX       4241457 Najee     Harris   Pit    
#>  8     2021               1      0 D/ST        -16025 49ers     D/ST     SF     
#>  9     2021               1      4 K          3055899 Harrison  Butker   KC     
#> 10     2021               1      4 BE         4035538 David     Montgom… Chi    
#> 11     2021               1      4 BE         4241985 J.K.      Dobbins  Bal    
#> 12     2021               1      4 BE           15880 Robert    Woods    LAR    
#> 13     2021               1      4 BE         4241389 CeeDee    Lamb     Dal    
#> 14     2021               1      4 BE         3915416 DJ        Moore    Car    
#> 15     2021               1      4 BE         3116165 Chris     Godwin   TB     
#> 16     2021               1      4 BE         2577327 Tyler     Lockett  Sea    
#> # … with 7 more variables: position <fct>, injuryStatus <chr>,
#> #   projectedScore <dbl>, actualScore <dbl>, percentStarted <dbl>,
#> #   percentOwned <dbl>, percentChange <dbl>
#>