Fantasy team roster

The players on a team roster with position, injury status, and playing stats.

team_roster(lid = getOption("lid"), old = FALSE, ...)

Arguments

lid
ESPN League ID, defaults to getOption("lid"). Use options(lid = <League ID>) to set your league ID; put that line in your .Rprofile file to set the option at start up.

old
If FALSE (default), return only the current season's data. If TRUE, return a list with all prior seasons.

...
Additional API query arguments. Use "week" and "year" as a shorthand for "ScoringPeriodId" and "seasonId" respectively.

Value

A tibble (or list) of players on a roster.

Examples

team_roster(lid = 252353)

#> $AGUS
#> # A tibble: 16 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020    17 AGUS  QB    3139477 Patr… Maho… KC    QB    A       0      0  
#>  2  2020    17 AGUS  RB    4045163 Miles Sand… Phi   RB    Q       0      0  
#>  3  2020    17 AGUS  RB    3042519 Aaron Jones GB    RB    A      12.3   14.5
#>  4  2020    17 AGUS  WR    3925357 Calv… Ridl… Atl   WR    A      14.7    2.6
#>  5  2020    17 AGUS  WR    4046692 Chase Clay… Pit   WR    A       6.82  16.1
#>  6  2020    17 AGUS  TE    4036133 T.J.  Hock… Det   TE    A       7.50   2.5
#>  7  2020    17 AGUS  FX    2576434 Melv… Gord… Den   RB    A      13.4   17  
#>  8  2020    17 AGUS  DS     -16028 Wash… D/ST  Wsh   DS    A       6.15  10  
#>  9  2020    17 AGUS  PK    3915165 Rodr… Blan… Ind   PK    A       8.14   6  
#> 10  2020    17 AGUS  BE       2330 Tom   Brady TB    QB    A      21.3   30.3
#> 11  2020    17 AGUS  BE    3045147 James Conn… Pit   RB    A       6.94   6.2
#> 12  2020    17 AGUS  BE    2976316 Mich… Thom… NO    WR    A       0      0  
#> 13  2020    17 AGUS  BE    2577327 Tyler Lock… Sea   WR    A       8.20  21  
#> 14  2020    17 AGUS  BE    3054212 Jonnu Smith Ten   TE    A       6.43   0.8
#> 15  2020    17 AGUS  BE    3916945 Dari… Slay… NYG   WR    A       6.72   2.2
#> 16  2020    17 AGUS  BE      15920 Lata… Murr… NO    RB    A       0      0  
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#> 
#> $PEPE
#> # A tibble: 16 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020    17 PEPE  QB    3917315 Kyler Murr… Ari   QB    Q      20.1   3.78
#>  2  2020    17 PEPE  RB    3068267 Aust… Ekel… LAC   RB    Q      13.5  11.1 
#>  3  2020    17 PEPE  RB    3919596 Chris Cars… Sea   RB    A      12.5   8.3 
#>  4  2020    17 PEPE  WR    2976212 Stef… Diggs Buf   WR    A       8.52  7.6 
#>  5  2020    17 PEPE  WR    4262921 Just… Jeff… Min   WR    A      12.4  13.3 
#>  6  2020    17 PEPE  TE    3121023 Dall… Goed… Phi   TE    Q       0     0   
#>  7  2020    17 PEPE  FX    4259545 D'An… Swift Det   RB    A      15.6  13   
#>  8  2020    17 PEPE  DS     -16022 Card… D/ST  Ari   DS    A       5.46  7   
#>  9  2020    17 PEPE  PK    3124679 Jason Sand… Mia   PK    A       8.79  9   
#> 10  2020    17 PEPE  BE    3120348 JuJu  Smit… Pit   WR    A       6.56 12.5 
#> 11  2020    17 PEPE  BE      15880 Robe… Woods LAR   WR    A       8.80  3.6 
#> 12  2020    17 PEPE  BE    4242214 Clyde Edwa… KC    RB    Q       0     0   
#> 13  2020    17 PEPE  BE    3116165 Chris Godw… TB    WR    A       8.99 25.3 
#> 14  2020    17 PEPE  BE    3045144 Tyler Boyd  Cin   WR    A       6.99  1.2 
#> 15  2020    17 PEPE  BE    2468609 Tays… Hill  NO    QB    A       6.65 10.6 
#> 16  2020    17 PEPE  BE     -16014 Rams  D/ST  LAR   DS    A       4.04 21.5 
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#> 
#> $BILL
#> # A tibble: 16 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020    17 BILL  QB      14881 Russ… Wils… Sea   QB    A      19.8   18.1
#>  2  2020    17 BILL  RB    3116593 Dalv… Cook  Min   RB    A       0      0  
#>  3  2020    17 BILL  RB    4360294 Anto… Gibs… Wsh   RB    Q      12.0    8.1
#>  4  2020    17 BILL  WR      16799 Allen Robi… Chi   WR    A      10.0    3.7
#>  5  2020    17 BILL  WR    4047650 DK    Metc… Sea   WR    A      11.2    2.1
#>  6  2020    17 BILL  TE      13229 Rob   Gron… TB    TE    A       6.50   2.9
#>  7  2020    17 BILL  FX    3915416 DJ    Moore Car   WR    A       8.95  10.1
#>  8  2020    17 BILL  DS     -16015 Dolp… D/ST  Mia   DS    A       4.43  -3.5
#>  9  2020    17 BILL  PK    3049899 Youn… Koo   Atl   PK    A       8.69   9  
#> 10  2020    17 BILL  BE    2576414 Rahe… Most… SF    RB    I       0      0  
#> 11  2020    17 BILL  BE    3121422 Terry McLa… Wsh   WR    Q       9.60  10  
#> 12  2020    17 BILL  BE    4036131 Noah  Fant  Den   TE    A       7.06   4.8
#> 13  2020    17 BILL  BE    4052042 James Robi… Jax   RB    Q       0      0  
#> 14  2020    17 BILL  BE    3932905 Dion… John… Pit   WR    A       7.29   9.6
#> 15  2020    17 BILL  BE     -16005 Brow… D/ST  Cle   DS    A       6.16  -1.5
#> 16  2020    17 BILL  BE     -16018 Sain… D/ST  NO    DS    A       4.70  17.5
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#> 
#> $CART
#> # A tibble: 16 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020    17 CART  QB    4038941 Just… Herb… LAC   QB    A      19.5   33.0
#>  2  2020    17 CART  RB    4047365 Josh  Jaco… LV    RB    A      13.4   20.9
#>  3  2020    17 CART  RB    2979843 Keny… Drake Ari   RB    A      11.0    5.1
#>  4  2020    17 CART  WR      16800 Dava… Adams GB    WR    A      16.0   10.6
#>  5  2020    17 CART  WR      16460 Adam  Thie… Min   WR    A      11.6    5.7
#>  6  2020    17 CART  TE      16813 Logan Thom… Wsh   TE    A       8.49   9.7
#>  7  2020    17 CART  FX    3122976 Jeff  Wils… SF    RB    A      14.3   20.8
#>  8  2020    17 CART  DS     -16033 Rave… D/ST  Bal   DS    A       6.38   6  
#>  9  2020    17 CART  PK    3055899 Harr… Butk… KC    PK    A       7.76   3  
#> 10  2020    17 CART  BE    2974858 Kenny Goll… Det   WR    Q       0      0  
#> 11  2020    17 CART  BE    3116385 Joe   Mixon Cin   RB    I       0      0  
#> 12  2020    17 CART  BE    2574808 Robby Ande… Car   WR    A       8.44   4  
#> 13  2020    17 CART  BE      16731 Bran… Cooks Hou   WR    A      10.1   28.6
#> 14  2020    17 CART  BE     -16026 Seah… D/ST  Sea   DS    A       5.77   2.5
#> 15  2020    17 CART  BE    3122866 Devo… Book… LV    RB    A       4.81   1.6
#> 16  2020    17 CART  BE      13934 Anto… Brown TB    WR    A       7.72  25.8
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#> 
#> $KIER
#> # A tibble: 16 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020    17 KIER  QB      14876 Ryan  Tann… Ten   QB    A      18.5   28.4
#>  2  2020    17 KIER  RB    3054850 Alvin Kama… NO    RB    A       0      0  
#>  3  2020    17 KIER  RB    4242335 Jona… Tayl… Ind   RB    A      17.3   37.4
#>  4  2020    17 KIER  WR      15795 DeAn… Hopk… Ari   WR    A       9.34   3.5
#>  5  2020    17 KIER  WR    2977187 Coop… Kupp  LAR   WR    A       0      0  
#>  6  2020    17 KIER  TE      15847 Trav… Kelce KC    TE    A       4.20   0  
#>  7  2020    17 KIER  FX    4035538 David Mont… Chi   RB    A      14.2   19.2
#>  8  2020    17 KIER  DS     -16002 Bills D/ST  Buf   DS    A       5.32  19.5
#>  9  2020    17 KIER  PK      15683 Just… Tuck… Bal   PK    A       8.92   8  
#> 10  2020    17 KIER  BE    2508176 David John… Hou   RB    A      12.1   18  
#> 11  2020    17 KIER  BE    3912550 Rona… Jone… TB    RB    A      12.4   13.8
#> 12  2020    17 KIER  BE    3025433 Mike  Davis Car   RB    Q       0      0  
#> 13  2020    17 KIER  BE    3886818 Myles Gask… Mia   RB    A      12.8   13.7
#> 14  2020    17 KIER  BE    3139522 Trav… Fulg… Phi   WR    A       2.83   1.5
#> 15  2020    17 KIER  BE    3128774 Kalen Ball… LAC   RB    A       5.31   9.6
#> 16  2020    17 KIER  BE      11237 Matt  Ryan  Atl   QB    A      18.6   22.7
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#> 
#> $CORE
#> # A tibble: 16 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020    17 CORE  QB    3916387 Lamar Jack… Bal   QB    A      22.1   24.2
#>  2  2020    17 CORE  RB    3051926 Gus   Edwa… Bal   RB    A       8.06   7  
#>  3  2020    17 CORE  RB    3119195 Chase Edmo… Ari   RB    A       7.47   2.5
#>  4  2020    17 CORE  WR    4241372 Marq… Brown Bal   WR    A       7.98  16.1
#>  5  2020    17 CORE  WR      15818 Keen… Allen LAC   WR    A       0      0  
#>  6  2020    17 CORE  TE    3116365 Mark  Andr… Bal   TE    A       8.53   2.7
#>  7  2020    17 CORE  FX      15072 Marv… Jone… Det   WR    A       8.69  30  
#>  8  2020    17 CORE  DS     -16003 Bears D/ST  Chi   DS    A       2.50  -2.5
#>  9  2020    17 CORE  PK    2971573 Ka'i… Fair… Hou   PK    A       7.70  17  
#> 10  2020    17 CORE  BE    3117251 Chri… McCa… Car   RB    Q       0      0  
#> 11  2020    17 CORE  BE    2576623 DeVa… Park… Mia   WR    A       7.57  11.6
#> 12  2020    17 CORE  BE    3051876 Evan  Engr… NYG   TE    A       5.79   1.7
#> 13  2020    17 CORE  BE    3052876 Will  Full… Hou   WR    A       0      0  
#> 14  2020    17 CORE  BE      16790 Jarv… Land… Cle   WR    A       8.51  11.4
#> 15  2020    17 CORE  BE     -16024 Char… D/ST  LAC   DS    A       4.68   4  
#> 16  2020    17 CORE  BE      16757 Derek Carr  LV    QB    A      17.3   20.9
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#> 
#> $NICK
#> # A tibble: 17 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020    17 NICK  QB    3122840 Desh… Wats… Hou   QB    A      23.7   25.8
#>  2  2020    17 NICK  RB    3043078 Derr… Henry Ten   RB    A      20.0   35  
#>  3  2020    17 NICK  RB    3045127 Wayne Gall… NYG   RB    A      10.0    6.5
#>  4  2020    17 NICK  WR    4360438 Bran… Aiyuk SF    WR    Q       0      0  
#>  5  2020    17 NICK  WR    3042778 Corey Davis Ten   WR    A       8.10   3.9
#>  6  2020    17 NICK  TE    3043275 Aust… Hoop… Cle   TE    A       4.81   9.7
#>  7  2020    17 NICK  FX    4241389 CeeD… Lamb  Dal   WR    A       7.50   4.6
#>  8  2020    17 NICK  DS     -16023 Stee… D/ST  Pit   DS    A       3.51   0  
#>  9  2020    17 NICK  PK    2985659 Wil   Lutz  NO    PK    A       7.58  10  
#> 10  2020    17 NICK  BE    2976499 Amari Coop… Dal   WR    Q       7.66   4.1
#> 11  2020    17 NICK  BE      13982 Julio Jones Atl   WR    Q       0      0  
#> 12  2020    17 NICK  BE    2977644 Todd  Gurl… Atl   RB    A       3.59   3.5
#> 13  2020    17 NICK  BE    3059915 Kare… Hunt  Cle   RB    A       7.99   4.1
#> 14  2020    17 NICK  BE      15091 Randy Bull… Cin   PK    A       0      0  
#> 15  2020    17 NICK  BE      16732 Eric  Ebron Pit   TE    A       0      0  
#> 16  2020    17 NICK  BE    3916430 Nyhe… Hines Ind   RB    A       9.74   6.7
#> 17  2020    17 NICK  IR    3040151 Geor… Kitt… SF    TE    A       9.49   6.8
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#> 
#> $KYLE
#> # A tibble: 15 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020    17 KYLE  QB       8439 Aaron Rodg… GB    QB    A      21.2   26  
#>  2  2020    17 KYLE  RB    3128720 Nick  Chubb Cle   RB    A      15.2   16.8
#>  3  2020    17 KYLE  RB    3115364 Leon… Four… TB    RB    A       4.21   2.7
#>  4  2020    17 KYLE  WR      16737 Mike  Evans TB    WR    Q      10.9    4.6
#>  5  2020    17 KYLE  WR    4047646 A.J.  Brown Ten   WR    A      10.9   21.1
#>  6  2020    17 KYLE  TE    2576925 Darr… Wall… LV    TE    A      10.9   17.7
#>  7  2020    17 KYLE  FX    3116406 Tyre… Hill  KC    WR    Q       0      0  
#>  8  2020    17 KYLE  DS     -16027 Bucc… D/ST  TB    DS    A       4.96   1.5
#>  9  2020    17 KYLE  PK    3051909 Dani… Carl… LV    PK    A       8.81   6  
#> 10  2020    17 KYLE  BE    3051392 Ezek… Elli… Dal   RB    A      13.2   12.1
#> 11  2020    17 KYLE  BE    3918298 Josh  Allen Buf   QB    A      15.9   19.3
#> 12  2020    17 KYLE  BE    3924365 Hayd… Hurst Atl   TE    A       5.82   8.8
#> 13  2020    17 KYLE  BE     -16019 Gian… D/ST  NYG   DS    A       4.85   8  
#> 14  2020    17 KYLE  BE    3126486 Deebo Samu… SF    WR    Q       0      0  
#> 15  2020    17 KYLE  BE    4240021 Cam   Akers LAR   RB    A       9.65   6.6
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#>