Fantasy team roster

The players on a team roster with position, injury status, and playing stats.

team_roster(lid = getOption("lid"), old = FALSE, ...)

Arguments

lid
ESPN League ID, defaults to getOption("lid"). Use options(lid = <League ID>) to set your league ID; put that line in your .Rprofile file to set the option at start up.

old
If FALSE (default), return only the current season's data. If TRUE, return a list with all prior seasons.

...
Additional API query arguments. Use "week" and "year" as a shorthand for "ScoringPeriodId" and "seasonId" respectively.

Value

A tibble (or list) of players on a roster.

Examples

team_roster(lid = 252353)

#> $AGUS
#> # A tibble: 16 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020     8 AGUS  QB    3139477 Patr… Maho… KC    QB    A      22.3   36.6
#>  2  2020     8 AGUS  RB    3045147 James Conn… Pit   RB    A      13.6   12  
#>  3  2020     8 AGUS  RB    2576434 Melv… Gord… Den   RB    A      10.1    1.9
#>  4  2020     8 AGUS  WR    3925357 Calv… Ridl… Atl   WR    Q      10.3    4.2
#>  5  2020     8 AGUS  WR    2577327 Tyler Lock… Sea   WR    A      12.5    2.3
#>  6  2020     8 AGUS  TE    3054212 Jonnu Smith Ten   TE    A       6.34   2.9
#>  7  2020     8 AGUS  FX      15920 Lata… Murr… NO    RB    A       7.05   2.9
#>  8  2020     8 AGUS  DS     -16012 Chie… D/ST  KC    DS    A       6.59   5.5
#>  9  2020     8 AGUS  PK      16339 Bran… McMa… Den   PK    A       8.54   6  
#> 10  2020     8 AGUS  BE       2330 Tom   Brady TB    QB    A      19.3   NA  
#> 11  2020     8 AGUS  BE    4045163 Miles Sand… Phi   RB    O       0     NA  
#> 12  2020     8 AGUS  BE    2976316 Mich… Thom… NO    WR    O       0      0  
#> 13  2020     8 AGUS  BE    3042519 Aaron Jones GB    RB    O       0      0  
#> 14  2020     8 AGUS  BE    3916945 Dari… Slay… NYG   WR    A       7.50  NA  
#> 15  2020     8 AGUS  BE    4036133 T.J.  Hock… Det   TE    A       5.89   6.5
#> 16  2020     8 AGUS  BE    4046692 Chase Clay… Pit   WR    A       6.90   8.2
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#> 
#> $PEPE
#> # A tibble: 17 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020     8 PEPE  QB    2573079 Cars… Wentz Phi   QB    A      20.2   NA  
#>  2  2020     8 PEPE  RB    4242214 Clyde Edwa… KC    RB    A      12.3    3.1
#>  3  2020     8 PEPE  RB    2980453 Jama… Will… GB    RB    A      14.8   12.2
#>  4  2020     8 PEPE  WR    2976212 Stef… Diggs Buf   WR    A      10.1    9.2
#>  5  2020     8 PEPE  WR    3045144 Tyler Boyd  Cin   WR    A       8.81  13.6
#>  6  2020     8 PEPE  TE    2975674 Robe… Tony… GB    TE    A       4.72   7.9
#>  7  2020     8 PEPE  FX    4262921 Just… Jeff… Min   WR    A       8.50   2.6
#>  8  2020     8 PEPE  DS     -16027 Bucc… D/ST  TB    DS    A       7.12  NA  
#>  9  2020     8 PEPE  PK    3049899 Youn… Koo   Atl   PK    A       8.44  13  
#> 10  2020     8 PEPE  BE    3120348 JuJu  Smit… Pit   WR    A       7.19   6.7
#> 11  2020     8 PEPE  BE      15880 Robe… Woods LAR   WR    A       9.61  21.4
#> 12  2020     8 PEPE  BE    3116165 Chris Godw… TB    WR    O       0     NA  
#> 13  2020     8 PEPE  BE    3917315 Kyler Murr… Ari   QB    A       0     NA  
#> 14  2020     8 PEPE  BE    3919596 Chris Cars… Sea   RB    O       0      0  
#> 15  2020     8 PEPE  BE    4259545 D'An… Swift Det   RB    A       9.38   2.3
#> 16  2020     8 PEPE  BE      16777 Carl… Hyde  Sea   RB    O       0      0  
#> 17  2020     8 PEPE  IR    3068267 Aust… Ekel… LAC   RB    I       0      0  
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#> 
#> $BILL
#> # A tibble: 17 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020     8 BILL  QB      14881 Russ… Wils… Sea   QB    A      23.4   27.9
#>  2  2020     8 BILL  RB    3116593 Dalv… Cook  Min   RB    A      15.9   46.6
#>  3  2020     8 BILL  RB    3928925 JaMy… Hasty SF    RB    A       7.37   9.1
#>  4  2020     8 BILL  WR    3915416 DJ    Moore Car   WR    A       9.42   5.5
#>  5  2020     8 BILL  WR    4047650 DK    Metc… Sea   WR    A      11.7   27.2
#>  6  2020     8 BILL  TE      13229 Rob   Gron… TB    TE    A       8.78  NA  
#>  7  2020     8 BILL  FX    3932905 Dion… John… Pit   WR    A       8.22   0.6
#>  8  2020     8 BILL  DS     -16017 Patr… D/ST  NE    DS    A       3.55  -0.5
#>  9  2020     8 BILL  PK      15683 Just… Tuck… Bal   PK    A       8.70   8  
#> 10  2020     8 BILL  BE    3121422 Terry McLa… Wsh   WR    A       0     NA  
#> 11  2020     8 BILL  BE      16799 Allen Robi… Chi   WR    A      10.1   11.7
#> 12  2020     8 BILL  BE    4360294 Anto… Gibs… Wsh   RB    A       0     NA  
#> 13  2020     8 BILL  BE    4036131 Noah  Fant  Den   TE    A       6.15   4.1
#> 14  2020     8 BILL  BE    4052042 James Robi… Jax   RB    A       0     NA  
#> 15  2020     8 BILL  BE      13934 Anto… Brown TB    WR    S       0     NA  
#> 16  2020     8 BILL  BE    3119195 Chase Edmo… Ari   RB    A       0     NA  
#> 17  2020     8 BILL  IR    2576414 Rahe… Most… SF    RB    I       0      0  
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#> 
#> $CART
#> # A tibble: 16 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020     8 CART  QB    4038941 Just… Herb… LAC   QB    A      18.8   20.7
#>  2  2020     8 CART  RB    4047365 Josh  Jaco… LV    RB    A      15.4   12.8
#>  3  2020     8 CART  RB    3051439 Bost… Scott Phi   RB    A      12.7   NA  
#>  4  2020     8 CART  WR      16800 Dava… Adams GB    WR    A      14.1   23.3
#>  5  2020     8 CART  WR      16460 Adam  Thie… Min   WR    A      10.2    2.7
#>  6  2020     8 CART  TE    3046439 Hunt… Henry LAC   TE    A       7.32   2  
#>  7  2020     8 CART  FX    2574808 Robby Ande… Car   WR    A      10.7    4.8
#>  8  2020     8 CART  DS     -16021 Eagl… D/ST  Phi   DS    A       6.58  NA  
#>  9  2020     8 CART  PK      11122 Matt  Prat… Det   PK    A       8.71   2  
#> 10  2020     8 CART  BE    2979843 Keny… Drake Ari   RB    O       0     NA  
#> 11  2020     8 CART  BE    2974858 Kenny Goll… Det   WR    Q       9.39   0  
#> 12  2020     8 CART  BE    3116385 Joe   Mixon Cin   RB    O       0      0  
#> 13  2020     8 CART  BE     -16033 Rave… D/ST  Bal   DS    A       4.55   3.5
#> 14  2020     8 CART  BE    3116164 Mike  Gesi… Mia   TE    A       5.58   0.8
#> 15  2020     8 CART  BE      10636 Mason Cros… GB    PK    A       7.69   2  
#> 16  2020     8 CART  BE      12483 Matt… Staf… Det   QB    A      15.0   22.4
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#> 
#> $KIER
#> # A tibble: 16 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020     8 KIER  QB      14876 Ryan  Tann… Ten   QB    A      17.0   17.3
#>  2  2020     8 KIER  RB    3054850 Alvin Kama… NO    RB    A      16.1   13.8
#>  3  2020     8 KIER  RB    3025433 Mike  Davis Car   RB    A      13.6    7.7
#>  4  2020     8 KIER  WR    2977187 Coop… Kupp  LAR   WR    A       8.94  11  
#>  5  2020     8 KIER  WR    3139522 Trav… Fulg… Phi   WR    A       9.43  NA  
#>  6  2020     8 KIER  TE      15847 Trav… Kelce KC    TE    A      10.3   16.9
#>  7  2020     8 KIER  FX    4242335 Jona… Tayl… Ind   RB    A      12.4    3.1
#>  8  2020     8 KIER  DS     -16014 Rams  D/ST  LAR   DS    A       5.48   3.5
#>  9  2020     8 KIER  PK    3124084 Joey  Slye  Car   PK    A       9.53   5  
#> 10  2020     8 KIER  BE      15795 DeAn… Hopk… Ari   WR    A       0     NA  
#> 11  2020     8 KIER  BE    2508176 David John… Hou   RB    A       0     NA  
#> 12  2020     8 KIER  BE    4035538 David Mont… Chi   RB    A      12.2   10.1
#> 13  2020     8 KIER  BE    3912550 Rona… Jone… TB    RB    A      11.0   NA  
#> 14  2020     8 KIER  BE    4038524 Gard… Mins… Jax   QB    D       0     NA  
#> 15  2020     8 KIER  BE    3886818 Myles Gask… Mia   RB    A      12.2   10.3
#> 16  2020     8 KIER  BE      13983 A.J.  Green Cin   WR    A       8.05   1.9
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#> 
#> $CORE
#> # A tibble: 17 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020     8 CORE  QB    3916387 Lamar Jack… Bal   QB    A      20.6   14.8
#>  2  2020     8 CORE  RB    4039359 Darr… Hend… LAR   RB    Q      11.8    5.8
#>  3  2020     8 CORE  RB      15826 Giov… Bern… Cin   RB    A      12.7   19.8
#>  4  2020     8 CORE  WR    4241372 Marq… Brown Bal   WR    A       9.06   6.3
#>  5  2020     8 CORE  WR      15818 Keen… Allen LAC   WR    A      10.8   12.7
#>  6  2020     8 CORE  TE    3116365 Mark  Andr… Bal   TE    A       8.59   3.2
#>  7  2020     8 CORE  FX      15072 Marv… Jone… Det   WR    A       6.34  15.9
#>  8  2020     8 CORE  DS     -16002 Bills D/ST  Buf   DS    A       5.18   2  
#>  9  2020     8 CORE  PK    3055899 Harr… Butk… KC    PK    A       9.30   5  
#> 10  2020     8 CORE  BE    2576623 DeVa… Park… Mia   WR    A       7.32   6.3
#> 11  2020     8 CORE  BE    3051876 Evan  Engr… NYG   TE    A       5.63  NA  
#> 12  2020     8 CORE  BE    3052876 Will  Full… Hou   WR    A       0     NA  
#> 13  2020     8 CORE  BE      11237 Matt  Ryan  Atl   QB    A      17.8   17.9
#> 14  2020     8 CORE  BE      16725 Sammy Watk… KC    WR    O       0      0  
#> 15  2020     8 CORE  BE      16790 Jarv… Land… Cle   WR    A       6.67   5.2
#> 16  2020     8 CORE  BE    3116136 Just… Jack… LAC   RB    A       9.50  14.2
#> 17  2020     8 CORE  IR    3117251 Chri… McCa… Car   RB    I       0      0  
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#> 
#> $NICK
#> # A tibble: 16 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020     8 NICK  QB    3915511 Joe   Burr… Cin   QB    A      19.0   18.9
#>  2  2020     8 NICK  RB    3043078 Derr… Henry Ten   RB    A      21.0   17.2
#>  3  2020     8 NICK  RB    3059915 Kare… Hunt  Cle   RB    A      17.9    7.3
#>  4  2020     8 NICK  WR      13982 Julio Jones Atl   WR    A      10.4   13.7
#>  5  2020     8 NICK  WR    4239993 Tee   Higg… Cin   WR    A       8.57   8.4
#>  6  2020     8 NICK  TE    3040151 Geor… Kitt… SF    TE    Q      12.9    3.9
#>  7  2020     8 NICK  FX    2977644 Todd  Gurl… Atl   RB    A      14.2   10.6
#>  8  2020     8 NICK  DS     -16023 Stee… D/ST  Pit   DS    A       4.13  17.5
#>  9  2020     8 NICK  PK      15091 Randy Bull… Cin   PK    A       8.33   7  
#> 10  2020     8 NICK  BE    2976499 Amari Coop… Dal   WR    A       7.86  NA  
#> 11  2020     8 NICK  BE    3122840 Desh… Wats… Hou   QB    A       0     NA  
#> 12  2020     8 NICK  BE    4241389 CeeD… Lamb  Dal   WR    A       6.60  NA  
#> 13  2020     8 NICK  BE      15971 Rex   Burk… NE    RB    A       7.12   3.1
#> 14  2020     8 NICK  BE      15349 Cole  Beas… Buf   WR    A       6.11   2.4
#> 15  2020     8 NICK  BE    3910544 Josh… Kell… LAC   RB    A       9.39   2.1
#> 16  2020     8 NICK  BE      12537 Jared Cook  NO    TE    A       5.18   9.8
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#> 
#> $KYLE
#> # A tibble: 17 x 15
#>     year  week team  slot       id first last  pro   pos   status  proj score
#>    <int> <int> <fct> <fct>   <int> <chr> <chr> <fct> <fct> <chr>  <dbl> <dbl>
#>  1  2020     8 KYLE  QB       8439 Aaron Rodg… GB    QB    A      19.2   22.5
#>  2  2020     8 KYLE  RB    3051392 Ezek… Elli… Dal   RB    A      13.2   NA  
#>  3  2020     8 KYLE  RB    3115364 Leon… Four… TB    RB    A      10.5   NA  
#>  4  2020     8 KYLE  WR    4047646 A.J.  Brown Ten   WR    A       9.95   8.4
#>  5  2020     8 KYLE  WR    3116406 Tyre… Hill  KC    WR    A      11.6   21.8
#>  6  2020     8 KYLE  TE    2576925 Darr… Wall… LV    TE    A       8.76   2.8
#>  7  2020     8 KYLE  FX      15825 Le'V… Bell  KC    RB    A       8.24   3.8
#>  8  2020     8 KYLE  DS     -16011 Colts D/ST  Ind   DS    A       4.87  17  
#>  9  2020     8 KYLE  PK    3051909 Dani… Carl… LV    PK    A       7.14   9  
#> 10  2020     8 KYLE  BE      16737 Mike  Evans TB    WR    A       9.27  NA  
#> 11  2020     8 KYLE  BE    4241475 Henry Rugg… LV    WR    A       5.58   0.9
#> 12  2020     8 KYLE  BE    3918298 Josh  Allen Buf   QB    A      20.1   12.5
#> 13  2020     8 KYLE  BE    3924365 Hayd… Hurst Atl   TE    A       5.84   5.4
#> 14  2020     8 KYLE  BE    3925347 Dami… Harr… NE    RB    A       9.33  16.2
#> 15  2020     8 KYLE  BE     -16019 Gian… D/ST  NYG   DS    A       1.67  NA  
#> 16  2020     8 KYLE  BE    3042910 Rash… Higg… Cle   WR    A       5.10   1.4
#> 17  2020     8 KYLE  IR    3128720 Nick  Chubb Cle   RB    I       0      0  
#> # … with 3 more variables: start <dbl>, rost <dbl>, change <dbl>
#>