Fantasy team rosters

The roster of all teams in a league.

team_roster(
  leagueId = ffl_id(),
  leagueHistory = FALSE,
  scoringPeriodId = NULL,
  ...
)

Arguments

leagueId

Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

leagueHistory

logical; Should the leagueHistory version of the API be called? If TRUE, a list of results is returned, with one element for each historical year of the league.

scoringPeriodId

Integer week of NFL season. By default, NULL will use the current week (see ffl_week()). Scoring periods are always one week in length, whereas matchups might be longer.

...

Additional queries passed to httr::GET(). Arguments are converted to a named list and passed to query alongside view.

Value

A dataframe (or list) with league teams.

See also

Examples

team_roster(leagueId = "42654852", scoringPeriodId = 1)
#> $AUS
#> # A tibble: 16 × 16
#>    seasonId scoringPeriodId teamId abbrev lineupSlot playerId firstName lastName
#>       <int>           <int>  <int> <fct>  <fct>         <int> <chr>     <chr>   
#>  1     2023               1      1 AUS    QB          3139477 Patrick   Mahomes 
#>  2     2023               1      1 AUS    RB          3117251 Christian McCaffr…
#>  3     2023               1      1 AUS    RB          4430807 Bijan     Robinson
#>  4     2023               1      1 AUS    WR          2977187 Cooper    Kupp    
#>  5     2023               1      1 AUS    WR          4372016 Jaylen    Waddle  
#>  6     2023               1      1 AUS    TE          2576925 Darren    Waller  
#>  7     2023               1      1 AUS    FLEX        3916148 Tony      Pollard 
#>  8     2023               1      1 AUS    D/ST         -16018 Saints    D/ST    
#>  9     2023               1      1 AUS    K             15683 Justin    Tucker  
#> 10     2023               1      1 AUS    BE          4361370 Chris     Olave   
#> 11     2023               1      1 AUS    BE          4241478 DeVonta   Smith   
#> 12     2023               1      1 AUS    BE          4360238 Dameon    Pierce  
#> 13     2023               1      1 AUS    BE          3126486 Deebo     Samuel  
#> 14     2023               1      1 AUS    BE            15795 DeAndre   Hopkins 
#> 15     2023               1      1 AUS    BE          3116165 Chris     Godwin  
#> 16     2023               1      1 AUS    BE          4567048 Kenneth   Walker …
#> # ℹ 8 more variables: proTeam <fct>, position <fct>, injuryStatus <chr>,
#> #   projectedScore <dbl>, actualScore <dbl>, percentStarted <dbl>,
#> #   percentOwned <dbl>, percentChange <dbl>
#> 
#> $BOS
#> # A tibble: 16 × 16
#>    seasonId scoringPeriodId teamId abbrev lineupSlot playerId firstName lastName
#>       <int>           <int>  <int> <fct>  <fct>         <int> <chr>     <chr>   
#>  1     2023               1      2 BOS    QB          3916387 Lamar     Jackson 
#>  2     2023               1      2 BOS    RB          3043078 Derrick   Henry   
#>  3     2023               1      2 BOS    RB          4047365 Josh      Jacobs  
#>  4     2023               1      2 BOS    WR          4362628 Ja'Marr   Chase   
#>  5     2023               1      2 BOS    WR          3116406 Tyreek    Hill    
#>  6     2023               1      2 BOS    TE          3116365 Mark      Andrews 
#>  7     2023               1      2 BOS    FLEX        4241389 CeeDee    Lamb    
#>  8     2023               1      2 BOS    D/ST         -16023 Steelers  D/ST    
#>  9     2023               1      2 BOS    K           4360234 Evan      McPhers…
#> 10     2023               1      2 BOS    BE          3116385 Joe       Mixon   
#> 11     2023               1      2 BOS    BE          4047650 DK        Metcalf 
#> 12     2023               1      2 BOS    BE            15818 Keenan    Allen   
#> 13     2023               1      2 BOS    BE          4248528 Christian Watson  
#> 14     2023               1      2 BOS    BE          4427366 Breece    Hall    
#> 15     2023               1      2 BOS    BE          4045163 Miles     Sanders 
#> 16     2023               1      2 BOS    BE          4379399 James     Cook    
#> # ℹ 8 more variables: proTeam <fct>, position <fct>, injuryStatus <chr>,
#> #   projectedScore <dbl>, actualScore <dbl>, percentStarted <dbl>,
#> #   percentOwned <dbl>, percentChange <dbl>
#> 
#> $CHI
#> # A tibble: 16 × 16
#>    seasonId scoringPeriodId teamId abbrev lineupSlot playerId firstName lastName
#>       <int>           <int>  <int> <fct>  <fct>         <int> <chr>     <chr>   
#>  1     2023               1      3 CHI    QB          4040715 Jalen     Hurts   
#>  2     2023               1      3 CHI    RB          3929630 Saquon    Barkley 
#>  3     2023               1      3 CHI    RB          4239996 Travis    Etienne…
#>  4     2023               1      3 CHI    WR          4262921 Justin    Jeffers…
#>  5     2023               1      3 CHI    WR          4569618 Garrett   Wilson  
#>  6     2023               1      3 CHI    TE            15847 Travis    Kelce   
#>  7     2023               1      3 CHI    FLEX        4374302 Amon-Ra   St. Bro…
#>  8     2023               1      3 CHI    D/ST         -16025 49ers     D/ST    
#>  9     2023               1      3 CHI    K           3055899 Harrison  Butker  
#> 10     2023               1      3 CHI    BE          4429795 Jahmyr    Gibbs   
#> 11     2023               1      3 CHI    BE          3042519 Aaron     Jones   
#> 12     2023               1      3 CHI    BE          3915511 Joe       Burrow  
#> 13     2023               1      3 CHI    BE          2976499 Amari     Cooper  
#> 14     2023               1      3 CHI    BE          4697815 Rachaad   White   
#> 15     2023               1      3 CHI    BE          3054850 Alvin     Kamara  
#> 16     2023               1      3 CHI    BE          4038941 Justin    Herbert 
#> # ℹ 8 more variables: proTeam <fct>, position <fct>, injuryStatus <chr>,
#> #   projectedScore <dbl>, actualScore <dbl>, percentStarted <dbl>,
#> #   percentOwned <dbl>, percentChange <dbl>
#> 
#> $DEN
#> # A tibble: 16 × 16
#>    seasonId scoringPeriodId teamId abbrev lineupSlot playerId firstName lastName
#>       <int>           <int>  <int> <fct>  <fct>         <int> <chr>     <chr>   
#>  1     2023               1      4 DEN    QB          3918298 Josh      Allen   
#>  2     2023               1      4 DEN    RB          3068267 Austin    Ekeler  
#>  3     2023               1      4 DEN    RB          3128720 Nick      Chubb   
#>  4     2023               1      4 DEN    WR            16800 Davante   Adams   
#>  5     2023               1      4 DEN    WR          2976212 Stefon    Diggs   
#>  6     2023               1      4 DEN    TE          4036133 T.J.      Hockens…
#>  7     2023               1      4 DEN    FLEX        4047646 A.J.      Brown   
#>  8     2023               1      4 DEN    D/ST         -16002 Bills     D/ST    
#>  9     2023               1      4 DEN    K           3051909 Daniel    Carlson 
#> 10     2023               1      4 DEN    BE          4239993 Tee       Higgins 
#> 11     2023               1      4 DEN    BE          4241457 Najee     Harris  
#> 12     2023               1      4 DEN    BE          4569173 Rhamondre Stevens…
#> 13     2023               1      4 DEN    BE          3925357 Calvin    Ridley  
#> 14     2023               1      4 DEN    BE          3932905 Diontae   Johnson 
#> 15     2023               1      4 DEN    BE          3045147 James     Conner  
#> 16     2023               1      4 DEN    BE          4048244 Alexander Mattison
#> # ℹ 8 more variables: proTeam <fct>, position <fct>, injuryStatus <chr>,
#> #   projectedScore <dbl>, actualScore <dbl>, percentStarted <dbl>,
#> #   percentOwned <dbl>, percentChange <dbl>
#>