Fantasy matchup scores

The score of each team in a matchup period and the match outcome.

tidy_scores(leagueId = ffl_id(), leagueHistory = FALSE, ...)

Arguments

leagueId
Numeric league ID or ESPN fantasy page URL. Defaults to getOption("fflr.leagueId"). Function fails if no ID is found.

leagueHistory
logical; Should the leagueHistory version of the API be called? If TRUE, a list of results is returned, with one element for each historical year of the league.

...
Additional queries passed to httr::GET() (e.g., scoringPeriodId). Arguments are converted to a named list to be passed alongside view.

Value

A tidy data frame of scores by team and matchup period.

Details

powerWins are added by comparing a team score versus every other team during that matchup period; the highest score would have won against all teams and the lowest score would have zero wins.

Examples

tidy_scores(leagueId = "42654852")

#> # A tibble: 64 × 9
#>    seasonId    id matchupPeriodId teamId abbrev isHome totalPoints isWinner
#>       <int> <int> <fct>            <int> <fct>  <lgl>        <dbl> <lgl>   
#>  1     2021     1 1                    1 AUS    TRUE         118.  FALSE   
#>  2     2021     1 1                    3 CHI    FALSE        119.  TRUE    
#>  3     2021     2 1                    2 BOS    TRUE         164.  TRUE    
#>  4     2021     2 1                    4 DEN    FALSE        135.  FALSE   
#>  5     2021     3 2                    3 CHI    TRUE         155.  FALSE   
#>  6     2021     3 2                    4 DEN    FALSE        160.  TRUE    
#>  7     2021     4 2                    1 AUS    TRUE         140.  TRUE    
#>  8     2021     4 2                    2 BOS    FALSE         93.7 FALSE   
#>  9     2021     5 3                    4 DEN    TRUE           1   FALSE   
#> 10     2021     5 3                    1 AUS    FALSE          0   TRUE    
#> # … with 54 more rows, and 1 more variable: powerWins <dbl>

tidy_scores(leagueId = "252353", leagueHistory = TRUE)

#> [[1]]
#> # A tibble: 120 × 9
#>    seasonId    id matchupPeriodId teamId abbrev isHome totalPoints isWinner
#>       <int> <int> <fct>            <int> <fct>  <lgl>        <dbl> <lgl>   
#>  1     2015     0 1                    7 GRIZ   TRUE            99 FALSE   
#>  2     2015     0 1                    4 BILL   FALSE          111 TRUE    
#>  3     2015     1 1                    5 BREW   TRUE            74 FALSE   
#>  4     2015     1 1                    8 CHAR   FALSE          107 TRUE    
#>  5     2015     2 1                    2 LOVE   TRUE           117 TRUE    
#>  6     2015     2 1                    1 GUS    FALSE          100 FALSE   
#>  7     2015     3 1                    6 K5     TRUE           111 TRUE    
#>  8     2015     3 1                    3 PEPE   FALSE           74 FALSE   
#>  9     2015     4 2                    4 BILL   TRUE            67 FALSE   
#> 10     2015     4 2                    8 CHAR   FALSE           94 TRUE    
#> # … with 110 more rows, and 1 more variable: powerWins <dbl>
#> 
#> [[2]]
#> # A tibble: 150 × 9
#>    seasonId    id matchupPeriodId teamId abbrev isHome totalPoints isWinner
#>       <int> <int> <fct>            <int> <fct>  <lgl>        <dbl> <lgl>   
#>  1     2016     0 1                    3 PEPE   TRUE          76.8 FALSE   
#>  2     2016     0 1                   11 MOOD   FALSE         79.0 TRUE    
#>  3     2016     1 1                    8 CHAR   TRUE          94.0 FALSE   
#>  4     2016     1 1                   10 NICK   FALSE        121.  TRUE    
#>  5     2016     2 1                    9 TEAM   TRUE         117.  TRUE    
#>  6     2016     2 1                    6 K5     FALSE        106.  FALSE   
#>  7     2016     3 1                    4 BILL   TRUE         112.  TRUE    
#>  8     2016     3 1                    1 GUS    FALSE         98.8 FALSE   
#>  9     2016     4 1                    2 LOVE   TRUE          78.7 FALSE   
#> 10     2016     4 1                    5 BRUZ   FALSE         95.7 TRUE    
#> # … with 140 more rows, and 1 more variable: powerWins <dbl>
#> 
#> [[3]]
#> # A tibble: 140 × 9
#>    seasonId    id matchupPeriodId teamId abbrev isHome totalPoints isWinner
#>       <int> <int> <fct>            <int> <fct>  <lgl>        <dbl> <lgl>   
#>  1     2017     0 1                    3 PEPE   TRUE          91.9 FALSE   
#>  2     2017     0 1                   11 MOOD   FALSE         95   TRUE    
#>  3     2017     1 1                    8 CHAR   TRUE          84.1 FALSE   
#>  4     2017     1 1                   10 NICK   FALSE        111.  TRUE    
#>  5     2017     2 1                    9 COLN   TRUE          80.4 FALSE   
#>  6     2017     2 1                    6 KIER   FALSE         95.1 TRUE    
#>  7     2017     3 1                    4 BILL   TRUE          71.5 TRUE    
#>  8     2017     3 1                    1 GUS    FALSE         58.7 FALSE   
#>  9     2017     4 1                    2 LOVE   TRUE          63.6 FALSE   
#> 10     2017     4 1                    5 CART   FALSE         92.8 TRUE    
#> # … with 130 more rows, and 1 more variable: powerWins <dbl>
#> 
#> [[4]]
#> # A tibble: 140 × 9
#>    seasonId    id matchupPeriodId teamId abbrev isHome totalPoints isWinner
#>       <int> <int> <fct>            <int> <fct>  <lgl>        <dbl> <lgl>   
#>  1     2018     0 1                    3 PEPE   TRUE          73.0 FALSE   
#>  2     2018     0 1                   11 KYLE   FALSE         84.9 TRUE    
#>  3     2018     1 1                    8 CHAR   TRUE          85.4 TRUE    
#>  4     2018     1 1                   10 NICK   FALSE         77.3 FALSE   
#>  5     2018     2 1                    9 COLN   TRUE          94.0 FALSE   
#>  6     2018     2 1                    6 KIER   FALSE        119.  TRUE    
#>  7     2018     3 1                    4 BILL   TRUE         185.  TRUE    
#>  8     2018     3 1                    1 GUS    FALSE         63.5 FALSE   
#>  9     2018     4 1                    2 ROWN   TRUE         115.  TRUE    
#> 10     2018     4 1                    5 CART   FALSE        105.  FALSE   
#> # … with 130 more rows, and 1 more variable: powerWins <dbl>
#> 
#> [[5]]
#> # A tibble: 112 × 9
#>    seasonId    id matchupPeriodId teamId abbrev isHome totalPoints isWinner
#>       <int> <int> <fct>            <int> <fct>  <lgl>        <dbl> <lgl>   
#>  1     2019     1 1                    8 CORY   TRUE          83.1 FALSE   
#>  2     2019     1 1                    4 BILL   FALSE        120.  TRUE    
#>  3     2019     2 1                   10 NICK   TRUE         105.  FALSE   
#>  4     2019     2 1                    6 KIER   FALSE        116.  TRUE    
#>  5     2019     3 1                   11 KYLE   TRUE         105.  FALSE   
#>  6     2019     3 1                    5 CART   FALSE        106.  TRUE    
#>  7     2019     4 1                    3 PEPE   TRUE         123.  TRUE    
#>  8     2019     4 1                    1 ANGS   FALSE         78.2 FALSE   
#>  9     2019     5 2                    4 BILL   TRUE          82.7 FALSE   
#> 10     2019     5 2                    6 KIER   FALSE         84.8 TRUE    
#> # … with 102 more rows, and 1 more variable: powerWins <dbl>
#> 
#> [[6]]
#> # A tibble: 112 × 9
#>    seasonId    id matchupPeriodId teamId abbrev isHome totalPoints isWinner
#>       <int> <int> <fct>            <int> <fct>  <lgl>        <dbl> <lgl>   
#>  1     2020     1 1                    3 PEPE   TRUE         103.  TRUE    
#>  2     2020     1 1                    1 AGUS   FALSE         68.5 FALSE   
#>  3     2020     2 1                   10 NICK   TRUE          86.0 FALSE   
#>  4     2020     2 1                    4 BILL   FALSE        134.  TRUE    
#>  5     2020     3 1                    5 CART   TRUE         149.  TRUE    
#>  6     2020     3 1                   11 KYLE   FALSE         94.7 FALSE   
#>  7     2020     4 1                    8 CORE   TRUE         119.  TRUE    
#>  8     2020     4 1                    6 KIER   FALSE         99.7 FALSE   
#>  9     2020     5 2                    1 AGUS   TRUE         144.  TRUE    
#> 10     2020     5 2                    4 BILL   FALSE        110.  FALSE   
#> # … with 102 more rows, and 1 more variable: powerWins <dbl>
#>